% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anno_AnnoEset.R
\name{AnnoEset}
\alias{AnnoEset}
\title{Annotate EsetList via bioconductor}
\usage{
AnnoEset(eset, probeid = "ID")
}
\arguments{
\item{eset}{a EsetList object}

\item{probeid}{the colnames of probe id in fData.Default is "ID".}
}
\value{
an annotated EsetList object(fData)
}
\description{
Annotate EsetList via bioconductor
}
\examples{
## This is a simulative process and available only with CORRECT VARIABLES
eset2 <- AnnoEset(eset)
View(Biobase::fData(eset2))
}
\seealso{
\code{\link[clusterProfiler]{bitr}};\code{\link[dplyr]{left_join}};\code{\link[readxl]{read_xlsx}}
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
