% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anno_AnnotateStringPPI.R
\name{AnnotateStringPPI}
\alias{AnnotateStringPPI}
\title{Re-annotation for ppi network from STRING web}
\usage{
AnnotateStringPPI(ppi, anno.abnormal)
}
\arguments{
\item{ppi}{a data frame from "... as simple tabular text output" in string PPI network analysis}

\item{anno.abnormal}{a list with wrong annotation name and right annotation elements}
}
\description{
Re-annotation for ppi network from STRING web
}
\examples{
## This is a simulative process and available only with CORRECT VARIABLES
anno.abnormal <- list(
"ENSG00000248993" = "XXbac-BPG181M17.5",
"PTPN20A" = "PTPN20",
"LRRC33" = "NRROS",
"C10orf54"  = "VSIR",
"EMR1" = "ADGRE1",
"1-Mar"= "MARCH1",
"FAIM3"="FCMR",
"ADORA3"= "TMIGD3"
)
ppi2 <- AnnotateStringPPI(ppi,anno.abnormal)
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
