
\name{common.annot}
\alias{common.annot}
\title{latest annotation contains ENSEMBL and other information of human transcription.}
\description{
 human annotation database
}
\usage{
  data(common.annot)
}
\arguments{
}
\details{
 The version gtf file of common.annot comes from is "gencode.v28.chr_patch_hapl_scaff.annotation.gtf".
}
\value{
a data frame
}
\references{

}
\author{
  Weibin Huang
}
\note{

}
\seealso{

}
\examples{

}
\keyword{common.annot}

