
\name{is.all.na}
\alias{is.all.na}
\title{Test whether a vector consists with NA values only}
\description{
 Test whether a vector consists with NA values only.
}
\usage{
  is.all.na(vt)
}
\arguments{
  \item{vt}{a vector.}
}
\details{

}
\value{

}
\references{

}
\author{
  Weibin Huang<654751191@qq.com>
}
\note{

}
\seealso{

}
\examples{
 a <- c(NA,NA);is.all.na(a)
#[1] TRUE
 b <- c(NA,2);is.all.na(b)
#[1] FALSE
 c <- c(1,3);is.all.na(c)
#[1] FALSE
}
\keyword{is.all.na}




