
\name{is.one.na}
\alias{is.one.na}
\title{Test whether a vector consists with at least one NA values.}
\description{
 Test whether a vector consists with at least one NA values.
}
\usage{
  is.one.na(vt)
}
\arguments{
  \item{vt}{a vector.}
}
\details{

}
\value{

}
\references{

}
\author{
  Weibin Huang
}
\note{

}
\seealso{

}
\examples{
a <- c(NA,NA);is.one.na(a)
# [1] TRUE
b <- c(NA,2);is.one.na(b)
# [1] TRUE
c <- c(1,3);is.one.na(c)
# [1] FALSE
}
\keyword{is.one.na}

