
\name{load1}
\alias{load1}
\title{load local .rda files by a pattern}
\description{
load1 help load local .rda files by a pattern
}
\usage{
load1(pattern,path=".")
}
\arguments{
  \item{pattern}{a character of the pattern of target file}
}
\details{

}
\value{

}
\references{

}
\author{
  Weibin Huang<654751191@qq.com>
}
\note{

}
\seealso{

}
\examples{

}
\keyword{load1(pattern)}




