
\name{save.Fastcornet}
\alias{save.Fastcornet}
\title{Save the result of Fastcornet series function}
\description{
    Save the result of Fastcornet series function
}
\usage{
save.Fastcornet(object,names = "love")
}
\arguments{
  \item{object}{result of Fastcornet series}
  \item{names}{part of saved file name}
}
\details{

}
\value{

}
\references{

}
\author{
  Weibin Huang<654751191@qq.com>
}
\note{

}
\seealso{

}
\examples{

}

\keyword{save.Fastcornet}




