
\name{show.color}
\alias{show.color}
\title{preview color vector}
\description{
 preview color vector
}
\usage{
 show.color(mycolor)
}
\arguments{
  \item{mycolor}{color vector like "#8DD3C7" or other type that can be identified by R.}
}
\details{

}
\value{

}
\references{

}
\author{
  Weibin Huang<654751191@qq.com>
}
\note{

}
\seealso{

}
\examples{
 data(mycolor)
 show.color(mycolor)
}
\keyword{show.color}




