% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_FastWGCNA.R
\name{BeforeWGCNA}
\alias{BeforeWGCNA}
\title{Before WGCNA}
\usage{
BeforeWGCNA(expr.matrix, design, log.convert = T, mad.portion = 0.75,
  mad.min = 0.01)
}
\arguments{
\item{expr.matrix}{expression matrix}

\item{design}{a design object}

\item{log.convert}{whether do log2 scale for expression matrix}

\item{mad.portion}{a portion of data you want base on median absolute deviation(mad) filter}

\item{mad.min}{the lower cut-off of mad filter}
}
\description{
Test the number of genes selected in WGCNA
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
\keyword{BeforeWGCNA}
