
\name{genenames.stad}
\alias{genenames.stad}
\title{gene names from TCGA-STAD with available annotation.}
\description{
 gene names from TCGA-STAD with available annotation.
}
\usage{

}
\arguments{
}
\details{
 gene names from TCGA-STAD with available annotation.
}
\value{
 a character
}
\references{

}
\author{
  Weibin Huang
}
\note{

}
\seealso{

}
\examples{

}
\keyword{genenames.stad}

