
\name{ROC.optimized}
\alias{ROC.optimized}
\title{Common ROC optimized via AUC}
\description{
 common ROC optimized base on expr.matrix and design object via AUC.
}
\usage{
ROC.optimized(expr.matrix,
              design,
              contrast.col,
              contrast.control,
              parallel = F)
}
\arguments{
}
\details{

}
\value{

}
\references{

}
\author{
  Weibin Huang<654751191@qq.com>
}
\note{

}
\seealso{

}
\examples{

}
\keyword{ROC.optimized}




