
\name{rowScale}
\alias{rowScale}
\title{rowScale}
\description{
对矩阵按行进行归一化处理。log.convert是否对matrix进行log2的转换
}
\usage{
  rowScale(matrix,log.convert = T)
}
\arguments{
  \item{log.convert}{是否对matrix进行log2的转换}
}
\details{

}
\value{

}
\references{

}
\author{
  Weibin Huang
}
\note{

}
\seealso{

}
\examples{

}
\keyword{rowScale}

