
\name{select.eset}
\alias{select.eset}
\title{fast way to select eset}
\description{
 fast way to select eset.it is equal with eset[,select.id]
}
\usage{
  select.eset(eset,select.id)
}
\arguments{
  \item{vt}{a vector.}
}
\details{

}
\value{

}
\references{

}
\author{
  Weibin Huang<654751191@qq.com>
}
\note{

}
\seealso{

}
\examples{

}
\keyword{select.eset}




