% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_ceRNAData.R
\name{ceRNAData}
\alias{ceRNAData}
\title{Get lnRNA-miRNA/mRNA-miRNA interaction from starBase}
\usage{
ceRNAData(object, verbose = F)
}
\arguments{
\item{object}{a name list(lncRNA/mRNA) of lncRNA and mRNAs.ENSEMBL id or SYMBOL id are all supported}
}
\value{
a name list(lncRNA/mRNA) of their interactive miRNAs
}
\description{
Get lnRNA-miRNA/mRNA-miRNA interaction from starBase
}
\examples{
object <- list(mRNA = c("TP53","TP53","NOTEXIST"),
               lncRNA = c("MALAT1","MALAT1","NOTEXIST"))
ceRNA <- ceRNAData(object)
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
