% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_ceRNANet.R
\name{ceRNANet}
\alias{ceRNANet}
\title{Get ceRNA network from starBase 3.0 by giving miRNAs symbol}
\usage{
ceRNANet(assembly = "hg19", geneType = "lncRNA", ceRNA = "HOTAIR",
  miRNAnum = 2, family = "all", pval = 0.01, fdr = 0.01,
  pancancerNum = 0, verbose = T)
}
\arguments{
\item{assembly}{genome reference of a species.Default is homo spacies("hg19").See \href{http://starbase.sysu.edu.cn/}{starbase}}

\item{geneType}{"mRNA","lncRNA","pseudogene","sncRNA".See \href{http://starbase.sysu.edu.cn/}{starbase}}

\item{ceRNA}{If you set a lncRNA as ceRNA,you should set \code{geneType="lncRNA"}}

\item{miRNAnum}{Default is 2.See \href{http://starbase.sysu.edu.cn/}{starbase}}

\item{family}{miRNA family names.See \href{http://starbase.sysu.edu.cn/}{starbase}}

\item{pval}{Default is 0.01.See \href{http://starbase.sysu.edu.cn/}{starbase}}

\item{fdr}{Default is 0.01.See \href{http://starbase.sysu.edu.cn/}{starbase}}

\item{pancancerNum}{Default is 0.See \href{http://starbase.sysu.edu.cn/}{starbase}}

\item{verbose}{whether to show verbose report}
}
\value{
a data frame with ceRNA interaction information
}
\description{
Get ceRNA network by giving miRNAs symbol
}
\examples{
## Get specified miRNA based ceRNA network
ceRNA = c("ENSG00000228630","ENSG00000228630","FKDKEK","FKDKEK")
family = "all"
# Or:
ceRNA = c("ENSG00000228630","ENSG00000228630","FKDKEK","FKDKEK")
family = "ABCKD" # Here,family parameter would be corrected to "all"

## Get specified miRNA based ceRNA network
ceRNA <- "all"
family <-  c("miR-1252-5p","miR-200bc-3p/429")

## Get specified lncRNA-miRNA based ceRNA network
ceRNA = c("ENSG00000228630")
family <-  c("miR-1252-5p","miR-200bc-3p/429")

## Get all lncRNA based ceRNA network
ceRNA <- "all"
family <-  "all"
net_ceRNA <- ceRNANet(assembly = "hg19",
                      geneType="lncRNA",
                      ceRNA,
                      miRNAnum=2,
                      family,
                      pval=0.01,
                      fdr=0.01,
                      pancancerNum=0,
                      verbose = T)
}
\seealso{
\code{starbase}\url{http://starbase.sysu.edu.cn/}
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
