% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_SequenceMotif.R
\name{combn2}
\alias{combn2}
\title{get all arrage for given n and k}
\usage{
combn2(vt, n)
}
\arguments{
\item{vt}{a vector}

\item{n}{selected counts}
}
\value{
a vector of arrage
}
\description{
get all arrage for given n and k
}
\examples{
base = c("A","C","U","G")
base2 <- combn2(base,c(3:5,9))
base3 <- combn2(base,c(3:5))
all(base2 \%in\% base3)
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
