% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_compareROC.R
\name{compareROC}
\alias{compareROC}
\title{Produce time dependent KM-ROC curves}
\usage{
compareROC(data, i.values, merge.i.values, timecutoff, roc.type, color,
  output.name, return.plot)
}
\arguments{
\item{data}{a data frame with survival data like time and status}

\item{i.values}{c("PNI1","NIHrisk").colnames of markers}

\item{merge.i.values}{list(merge1 = i.values).a list of merge markers selection}

\item{timecutoff}{c(1-365,3-365,5-365).cut-off for time series}

\item{roc.type}{c("ggplot","pROC")[1].ggplot recommanded}

\item{color}{brewer.pal(12, "Set3")[c(1,3,4)].curve colors}

\item{output.name}{part of PDF output name}
}
\description{
compareROC can help produce a time dependent KM curve,such as 1-year,3-year or 5-year KM curves.
}
\examples{
data(fat)
# markers=c("BMI","Waist","WHR","belly fat thickness") #marker colnames
# status =  "outcomes" #survival status
# merge.markers = list(c("BMI","Waist"))# the markers list that you want to merge as a co-prognostic factor
# return.plot = T # if T，return plot;if F，return data
# roc.type = c("ggplot","pROC")[1] # ggplot style is recommanded
# title="this is a title" # plot title
# color = brewer.pal(12, "Set3")[1:5] #curve colors.The lenght of colors must >= the number of curves.
# output.name = "ROC of something" # part of PDF name
compareROC2(data=fat,
            markers,
            status,
            merge.markers=NULL,
            return.plot=T,
            roc.type,
            title,
            color,
            output.name="test1")
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
