% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_compareROC2.R
\name{compareROC2}
\alias{compareROC2}
\title{get ROC Curve of one or more variates for a binary status}
\usage{
compareROC2(data, markers, status, merge.markers = NULL,
  roc.type = "ggplot", title = "ROC Curves", color = NULL,
  half.border = F, reference = T, legend.position = c(0.8, 0.2),
  show.auc = T, auc.digits = 2, width = 10, height = 10,
  names = "love")
}
\arguments{
\item{data}{a data frame}

\item{markers}{target markers}

\item{status}{the dependent variable in ROC.like "N.status"}

\item{merge.markers}{A list of markers you want to merge.Default is NULL.The merge stragegy is based on Generalized Linear Models(glm)}

\item{roc.type}{one of "ggplot" and "pROC"."ggplot" is recommanded and the default setting}

\item{title}{the plot title}

\item{color}{Default is NULL.You can set other colors like "#8DD3C7"}

\item{half.border}{whether to show half border style}

\item{reference}{whether to show a reference line in ROC plot}

\item{legend.position}{the position of legend in ggplot}

\item{show.auc}{whether to show related auc in legend labels}

\item{auc.digits}{the digits of auc in legend labels}

\item{width, height}{the size of saved PDF plot}

\item{names}{part of saved PDF plot names}
}
\description{
compareROC2 helps get ROC curves of one or more variates for a binary status.It supports the merge of lots of variates via glm function.
}
\examples{
## This is a simulative process and NOT RUN
train.LM.ROC <- compareROC2(
 data = design.x,
 markers = mgenes.s,
 status ="lymphatic.metastasis",
 merge.markers=list(merge = mgenes.s),
 roc.type = "ggplot",
 title="ROC Curves of Lymphatic Metastasis",
 color = mycolor[c(1,20,3,5,6,7,4)],
 legend.position=c(0.8,0.2),
 show.auc = T,auc.digits = 2,
 width = 10,height = 10,
 names =paste0("seed",seed,"_train.mgenes")
)
## Quick Start
data(fat)
markers=c("BMI","Waist","WHR","belly fat thickness") #marker #'colnames
status =  "outcomes" #survival status
merge.markers = list(c("BMI","Waist"))# the markers list that you want to merge as a co-prognostic factor
return.plot = T # if T，return plot;if F，return data
roc.type = c("ggplot","pROC")[1] # ggplot style is recommanded
title="this is a title" # plot title
color = brewer.pal(12, "Set3")[1:5] #curve colors.The lenght of colors must >= the number of curves.
output.name = "ROC of something" # part of PDF name
compareROC2(data=fat,
           markers,
           status,
           merge.markers=NULL,
           return.plot=T,
           roc.type,
           title,
           color,
           output.name="test1")
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
