% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anno_convert.R
\name{convert}
\alias{convert}
\title{multiple annotation based on dataset common.annot.}
\usage{
convert(vt, fromtype = "ENSEMBL", totype = "SYMBOL",
  db = common.annot)
}
\arguments{
\item{vt}{genes you want to annotation}

\item{fromtype}{the raw type of the genes.}

\item{totype}{the annotated type of the genes.}

\item{db}{a dataset containing annotation information.In lucky package,it is often common.annot dataset.}
}
\description{
multiple annotation based on dataset common.annot.Its colnames include:"ENSEMBL","ENSEMBL.Versions","SYMBOL","GENENAME","ENTREZID","UNIGENE" and "gene_type".
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
