% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_CorrectTNM.R
\name{CorrectTNM}
\alias{CorrectTNM}
\title{Version control for TCGA TNM stage}
\usage{
CorrectTNM(data, version = "stageVesion", toVersion = 8,
  T.status = "pT", N.status = "pN", M.status = "M.status",
  Np.counts = "Np.count", correctFile = NULL)
}
\arguments{
\item{data}{design object}

\item{version}{the colname of AJCC stage version.Note that the element of version should be "6th" like.}

\item{toVersion}{the final version you want to change.}

\item{T.status}{the colnames of T stage}

\item{N.status}{the colnames of N stage}

\item{M.status}{the colnames of M stage}

\item{Np.counts}{the colnames of positive lymph nodes}

\item{correctFile}{the path of reference database.Default = \code{system.file("extdata", "STAD_AJCC stage.xlsx", package = "lucky")}.It supports self-defined dataset with same format.}
}
\value{
a data frame
}
\description{
Version control for TCGA TNM stage
}
\details{
The most part of correction is the data from \code{correctFile}. It should be kept latest.\cr
For gastric cancer,before 5th(including 4th),the nodal involvement is classified into pN0, pN1, and pN2, based on the site of the metastasis in relation to the primary tumor.Afterwards,the number of positive lymph nodes become the details to stage N.
}
\examples{
## get your design data
load("E:/RCloud/database/DataDownload/TCGA-TCGABiolinks/STAD_Download/2018-12-3/design_STAD.rda")
colnames(design)
data = design
## update the TNM stage to 8th version
data2 <- CorrectTNM(data,
                    version = "stageVesion",
                    toVersion = 8,
                    T.status = "pT",
                    Np.counts = "Np.count",
                    M.status = "M.status")
}
\seealso{
\href{https://link.springer.com/book/10.1007/978-3-642-82982-6}{UICC 4th}
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
