
\name{fpkm}
\alias{fpkm}
\title{a simple dataset of fpkm values}
\description{
 a simple dataset of fpkm values form TCGA-STAD
}
\usage{
}
\arguments{

}
\details{
 a data frame with patient cols and gene rows.
}
\value{
a data frame
}
\references{

}
\author{
  Weibin Huang
}
\note{

}
\seealso{
  FPKMtoTPM()
}
\examples{

}
\keyword{fpkm}

