
\name{journals}
\alias{journals}
\title{SCI journals dataset}
\description{
 journals dataset contains lots of information of medical SCI journals.
}
\usage{

}
\arguments{
 \item{letpubid}{the letpubids  of journals in letpub}
\item{journals}{the names of journals}
\item{abbr}{the abbreviation of journal names}
\item{issn}{the ISSN code of journals}
\item{IF}{the impact factor of journals}
\item{IF5}{the merge impact factor of journals in recent 5 years}
\item{selfcited}{self-cited ratio}
\item{website}{the website of the journal}
\item{OA}{whether a OA journal}
\item{research}{the research direction}
\item{area}{the area of the journal lik USA}
\item{regular}{the period of article checking}
\item{BasicSubject}{the basic subject of the journal}
\item{SecondSubject}{the second subject of the journal}
\item{TopJournal}{whether a top journal}
\item{ReviewJournal}{whether a review journal}
\item{speed}{the speed of article checking}
\item{recieve}{the ratio of article recieved}
}
\details{

}
\value{

}
\references{

}
\author{
  Weibin Huang<654751191@qq.com>
}
\note{

}
\seealso{

}
\examples{

}
\keyword{journals}




