% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_DESeq1.R
\name{DESeq1}
\alias{DESeq1}
\title{Standard process for DESeq2 RNA-Seq differencial expression analysis}
\usage{
DESeq1(counts, design, contrast.col, contrast.level = c("Control",
  "Treat"), contrast.control = "Control", count.filter = 10,
  cutoff.lFC = 1, cutoff.padj = 0.1, report = F, save.file = T,
  names = "love")
}
\arguments{
\item{counts}{raw counts matrix}

\item{design}{design object with sample rows and clinic feature cols}

\item{contrast.col}{the colnames of contrast.Like "N.status"}

\item{count.filter}{the rowmean filter of counts}

\item{cutoff.lFC}{the cutoff of log foldchange}

\item{cutoff.padj}{the cutoff of adjusted P value}

\item{report}{whether to do quality control report.Default is F.}

\item{save.file}{whether to save .rda object for DESeqList}

\item{names}{part name of saved files}
}
\value{
DESeqList
}
\description{
DESeq1 is a customed way to use DESeq2 package for differencially expression analysis based on raw counts and design object,and create a DESeqList with dds,difSig and siggenes object.
}
\examples{
library(lucky)
## data preparation
data("rna.counts")
data("rna.design")
counts = rna.counts;rm(rna.counts)
design = rna.design;rm(rna.design)

## DESeq2 package standard pipeline
dl1 <- DESeq1(counts,
              design,
              contrast.col= "condition",
              contrast.level =  c("normal","tumor"),
              contrast.control = "normal",
              count.filter=10,
              cutoff.lFC = 2,
              cutoff.padj = 0.05,
              save.file = F,
              names = "love",
              report = T)
View(dl1)
}
\seealso{
\code{\link["DESeq2"]{DESeqDataSetFromMatrix}};\code{\link["DESeq2"]{DESeq}};
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
