% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_exampleLassoCox.R
\name{exampleLassoCox}
\alias{exampleLassoCox}
\title{Visualize and analysis of cox.optimized2 result by giving a specified seed}
\usage{
exampleLassoCox(expr.matrix, design, select, event.status, event.time,
  event.lower, k = 10, seed, optimize.method = "min", plot.label = T,
  plot.lwd = 2, line.lty = 3, line.lwd = 2, line.col = "blue",
  verbose = T, save.file = T, names = "love")
}
\arguments{
\item{expr.matrix}{gene expression with sample cols and genes/markers rows.}

\item{design}{design object with characters cols and sample rows.}

\item{select}{intersting genes/markers}

\item{event.status}{a vector of event status names}

\item{event.time}{a vector of event time names}

\item{event.lower}{the cutoff(>) ofevent.time}

\item{k}{number of folds.Default is 10.}

\item{seed}{a good example from the result of \code{uniqueModel}}

\item{optimize.method}{Default is "min".You can also use "1se"}

\item{plot.label}{whether to show labels in the plot of glmnet object}

\item{plot.lwd}{the line width of the plot of glmnet object}

\item{line.lty}{the line type of the cutoff line in the plot of glmnet object}

\item{line.lwd}{the line width of the cutoff line in the plot of glmnet object}

\item{line.col}{the line color of the cutoff line in the plot of glmnet object}

\item{verbose}{whether to do a plot report}

\item{save.file}{whether to save plot}

\item{names}{part name of the saved file}
}
\description{
Visualize and analysis of cox.optimized2 result by giving a specified seed
}
\examples{
## This is a simulative process and available only with CORRECT VARIABLES
}
\seealso{
\code{\link[glmnet]{glmnet}};\code{\link[glmnet]{cv.glmnet}};
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
