% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anno_fa.annotate.R
\name{fa.annotate}
\alias{fa.annotate}
\title{annotate probe sequence via fasta object}
\usage{
fa.annotate(sequence, fa.annotation, logical = F)
}
\arguments{
\item{sequence}{one or multiple sequeces (like "AGCGC")}

\item{fa.annotation}{fa.annotation object from lucky::get.fa.annoatation()}

\item{logical}{see parrallel::detectCores()}
}
\value{
a vector containing ENSEMBL ids.
}
\description{
annotate probe sequence via fasta object
}
\examples{
ensembl <- fa.annotate(sequence,fa.annotation)
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
