% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_Fastboxplot.R
\name{Fastboxplot}
\alias{Fastboxplot}
\title{fast way to draw a boxplot based on ggplot2}
\usage{
Fastboxplot(data, x, y, fill, facet, palette = NULL, x.lab, y.lab,
  legend.position = "none", x.test.angle = 0, add.jitter = T)
}
\arguments{
\item{data}{a data frame}

\item{x, y, fill, facet}{parameters of ggpubr::ggboxplot}

\item{palette}{fill color of box}

\item{x.lab}{the title of x axis}

\item{y.lab}{the title of y axis}

\item{legend.position}{legend.position}

\item{x.test.angle}{the angle of text on the x axis}

\item{add.jitter}{whether add points with jitter style}
}
\value{
a boxplot
}
\description{
fast way to draw a boxplot
}
\examples{
data(mtcars)
data=mtcars
x="vs"
y="disp"
fill="vs"
facet=NULL
facet="am"
x.lab="vs of mtcars"
y.lab="disp of mtcars"

## quick start of Fastboxplot
Fastboxplot(data,
            x,y,fill,facet,
            palette=NULL,
            x.lab,y.lab,
            legend.position="none",
            x.test.angle=0,
            add.jitter=T)

#change fill color
show.color(mycolor)
Fastboxplot(data,
            x,y,fill,facet,
            palette=c("#8DD3C7","#FB8072"),
            x.lab,y.lab,
            legend.position="none",
            x.test.angle=0,
            add.jitter=T)
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
