% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_FastCeRNA_edge.R
\name{FastCeRNA_edge}
\alias{FastCeRNA_edge}
\title{Get miRNA-lncRNA-mRNA edge data after FastCeRNA}
\usage{
FastCeRNA_edge(object, lncRNA.col = "lncRNAs", pcRNA.col = "Genes",
  miRNA.col = "miRNAs", Cor.col = "cor")
}
\arguments{
\item{object}{the result of \code{\link{FastCeRNA}}}

\item{lncRNA.col}{colnames of lncRNA}

\item{pcRNA.col}{colnames of protein-coding RNA,}

\item{miRNA.col}{colnames of miRNA}

\item{Cor.col}{colnames of Pearson corelation}
}
\value{
a data frame of miRNA-lncRNA-mRNA edge data
}
\description{
Get miRNA-lncRNA-mRNA edge data after FastCeRNA
}
\examples{
## load data
load("E:/iProjects/LM_WGCNA/data/LM_WGCNA_01A/rda/lm_ceRNA.rda")
load("E:/iProjects/LM_WGCNA/data/LM_WGCNA_01A/rda/lm_moduleGenes.rda")

## filter
lm_ceRNA_hubModules <- dplyr::filter(lm_ceRNA,Genes \%in\% as.character(moduleGenes),hyperPValue <= 0.05)

## Get edge data
ceRNA_edge <- FastCeRNA_edge(object = lm_ceRNA_hubModules,
                             lncRNA.col = "lncRNAs",
                             pcRNA.col = "Genes",
                             miRNA.col = "miRNAs",
                             Cor.col = "cor")
## Or easier way
ceRNA_edge <- FastCeRNA_edge(lm_ceRNA_hubModules)
View(ceRNA_edge)
}
\seealso{
\code{\link{FastCeRNA}}.
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
