% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_FastCeRNA_edge2.R
\name{FastCeRNA_edge2}
\alias{FastCeRNA_edge2}
\title{a simple method for ceRNA network edge/node selection}
\usage{
FastCeRNA_edge2(object, lncRNA.col = "lncRNAs", pcRNA.col = "Genes",
  miRNA.col = "miRNAs", pmcor = "pmcor", pmcor.Pvalue = "corPValue",
  lmcor = "lmcor", plcor = "plcor", cutoff = list(pl = 0.5, lm = 0.2,
  pm = 0.2), filter = T, verbose = T)
}
\arguments{
\item{object}{the result of \code{\link{FastCeRNA}}}

\item{lncRNA.col}{colnames of lncRNA}

\item{pcRNA.col}{colnames of protein-coding RNA,}

\item{miRNA.col}{colnames of miRNA}

\item{pmcor}{the col of protein-coding and miRNA correlation}

\item{pmcor.Pvalue}{the p value of pmcor}

\item{lmcor}{the col of lncRNA and miRNA correlation}

\item{plcor}{the col of protein-coding and miRNA correlation}

\item{cutoff}{the cutoff of correlation.Default is \code{list(pl=0.5,lm=0.2,pm=0.2)}}

\item{filter}{whether to do filter}

\item{verbose}{whether do verbose report}
}
\value{
a data frame
}
\description{
a simple method for ceRNA network edge/node selection
}
\details{
FastCeRNA_edge2 select ceRNA network:positive lnc-miRNA correaltion;miR negative.
}
\examples{
## This is a simulative process and available only with CORRECT VARIABLES
load("E:/iProjects/LM_WGCNA/data/LM_WGCNA_02C/rda/lm_ceRNA.rda")
df <- FastCeRNA_edge2(object)

## other parameters
colnames(lm_ceRNA)
object = lm_ceRNA
lncRNA.col = "lncRNAs"
pcRNA.col = "Genes"
miRNA.col = "miRNAs"
pmcor = "pmcor"
lmcor="lmcor"
plcor = "plcor"
filter = T
verbose = T
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
