% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_Fastcornet2.R
\name{Fastcornet2}
\alias{Fastcornet2}
\title{Draw Network plot of D3 style}
\usage{
Fastcornet2(data, source.col = NULL, target.col = NULL, value = NULL,
  color.group = list(), control.group = NULL, ratio = 1,
  legend = T, size = 15, linkColour = NULL,
  linkColour.type = list(name = "Blues", select.lower = 1, select.upper =
  9), control.linkColour = NULL, colourScale.type = NULL,
  nodeColour = list(), linkDistance = 300, linkWidth = 6,
  opacity = 1, arrows = F, opacityNoHover = 0, zoom = T)
}
\arguments{
\item{data}{a data frame with source,target and value cols}

\item{source.col}{the colnames of source}

\item{target.col}{the colnames of target}

\item{value}{the colnames of value}

\item{color.group}{a list of elements goup.}

\item{control.group}{Default is NULL.The group you want highlight.}

\item{ratio}{The portion of data you select}

\item{legend}{Whether show lengend}

\item{size}{a numeric to adjust the size of nodes}

\item{linkColour}{the link color.If it's NULL,the link color would be provided by \code{linkColour.type} parameter.}

\item{linkColour.type}{a list of link color for \code{value} mapping.}

\item{control.linkColour}{the color of highlight group.It only works when \code{control.group} exists.}

\item{colourScale.type}{d3-scale-chromatic style object.\cr
1="d3.schemeCategory20",\cr
2="d3.schemeCategory20b",\cr
3="d3.schemeCategory20c",\cr
4="d3.schemeCategory10".\cr
Default is 1}

\item{nodeColour}{the named list of group color.It is unavailable when colourScale.type exists}

\item{linkDistance}{link distance}

\item{linkWidth}{the width of link.If it is not a numeric,the default \code{"JS(function(d) { return Math.sqrt(d.value); })" } would be used.}

\item{opacity}{numeric value of the proportion opaque you would like the graph elements to be}

\item{arrows}{Whether show direction between two nodes}

\item{opacityNoHover}{the legend intensity.From 0-1.Default is 0}

\item{zoom}{logical value to enable (TRUE) or disable (FALSE)}
}
\value{
a list contain a network plot information and edge/node data.
}
\description{
Draw Network plot of D3 style
}
\details{
linkColour.type make full use of RColorBrewer and would be a good color strategy provider.The parameters of linkColour.type is including:\cr
1.names: One of "Blues","BuGn","BuPu","GnBu","Greens","Greys","Oranges","OrRd","PuBu","PuBuGn","PuRd","Purples","RdPu","Reds" ,"YlGn","YlGnBu","YlOrBr" and "YlOrRd";\cr
2.select.lower: the lower color order;\cr
3.select.upper: the upper color order.
}
\examples{
library(lucky)
data=state.x77;colnames(data)
## get MIC cormatrix
result1 <- FastMINE(data,
                    transposition = F,
                    method = "all.pairs",
                    control.markers="Income",
                    target.markers=NULL)

## long data frame network
data <- result1[["MINE.result"]][,1:3]
colnames(data) #[1] "X var"  "Y var" "MIC (strength)"
source.col="X var";target.col="Y var";value="MIC (strength)"

## parameters of Fastcornet2
color.group = list(
  love = c("Murder","Frost","Area","HS Grad"),
  hate = c("Life Exp","Illiteracy"),
  control = c("Population","Income"))
control.group = "control"
linkColour.type = list(
  name =  "Blues",
  select.lower = 1,
  select.upper = 3
)

# control.linkColour = "red"
control.linkColour = NULL
colourScale.type = NULL
nodeColour=list(
  love = mycolor[6],
  hate = mycolor[8],
  control = mycolor[35])

## Quick Start
nw2 <- Fastcornet2(data,
                   source.col,
                   target.col,
                   value,
                   color.group,
                   control.group,
                   ratio=0.8,
                   legend=T,
                   size = 15,
                   linkColour=NULL,
                   linkColour.type,
                   control.linkColour,
                   colourScale.type = NULL,
                   nodeColour);nw2$netplot
}
\seealso{
\code{\link[networkD3]{forceNetwork}}
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
