% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_FastEnrichPlot.R
\name{FastEnrichPlot}
\alias{FastEnrichPlot}
\title{Fast way to draw enrich plot after clusterProfilter enrichfunction}
\usage{
FastEnrichPlot(object, pvalueCutoff = NULL, qvalueCutoff = NULL,
  id.col = "Description", select = NULL,
  x.title = "Biological Process terms", y.title = "-log10(FDR)",
  size = 15, short.cutoff = 46, topshow = 10)
}
\arguments{
\item{object}{result from \code{\link[clusterProfilter]{enrichGO}} or \code{\link[clusterProfilter]{enrichKEGG}} or \code{\link[clusterProfilter]{gseGO}} or \code{\link[clusterProfilter]{enrichKEGG}}}

\item{pvalueCutoff}{p value cutoff}

\item{qvalueCutoff}{q value cutoff}

\item{id.col}{ID you want to show in final plot}

\item{select}{if \code{select = NULL},the number of \code{topshow} would be shown.If \code{select != NULL},selected terms would be showed and thus \code{topshow} is not effecient.}

\item{x.title}{the title of x axis}

\item{y.title}{the title of y axis}

\item{size}{the size of whole plot}

\item{short.cutoff}{if the number of character of terms is larger than \code{short.cutoff},it would automatically change row in the showed plot.Defaut is 46.If you don't want change-row,just set \code{short.cutoff = Inf}}

\item{topshow}{the number of top iterms ordered by FDR you want to show.It works only \code{select = NULL}}
}
\value{
LuckyList Object
}
\description{
Fast way to draw enrich plot after clusterProfilter enrichfunction
}
\examples{
## This is a simulative process and available only with CORRECT VARIABLES
load("E:/iProjects/LM_WGCNA/data/LM_WGCNA_02A/rda/goList.rda")
object <- goList$GOEnrichment$BP
res_EP <- FastEnrichPlot(object = object,
                         pvalueCutoff = NULL,
                         qvalueCutoff = NULL,
                         id.col = "Description",
                         select = NULL,
                         x.title = "Biological Process terms",
                         y.title = "-log10(FDR)",
                         size = 15,
                         short.cutoff = 46,
                         topshow = 10)
win.graph(12,9);res_EP$Plot

pvalueCutoff = 0.05
qvalueCutoff = 0.05
select = res2$ID[1:10]
id.col = "Description"
x.title = "Biological Process terms"
y.title = "-log10(FDR)"
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
