% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_FastGO.R
\name{FastGO}
\alias{FastGO}
\title{GO analysis via clusterProfiler package}
\usage{
FastGO(genes, geneList, default.universe = F, classlevel = 2:2,
  OrgDb = NULL, keyType = NULL, pAdjustMethod = "BH",
  pvalueCutoff = 0.05, qvalueCutoff = 0.05, cnet.showCategory = 5,
  verbose = TRUE, save.path = "GO", names = "love")
}
\arguments{
\item{genes}{a character of gene id.}

\item{geneList}{a numeric with id names.For example,logFoldChang with ENSEMBL id names.Note that the names of \code{geneList} must be the same type of \code{genes}}

\item{default.universe}{whether to use default universe.If \code{default.universe = F},the background genes would be provide by \code{geneList}}

\item{classlevel}{Default is 2.I had tested that 2:7 was still available.If many levels is set,the process is time-consuming}

\item{OrgDb}{OrgDb dataset.If NULL,use "org.Hs.eg.db"}

\item{keyType}{the type of gene.Support automatically test}

\item{pAdjustMethod}{the mathod of pvalue adjustment}

\item{pvalueCutoff}{a list of pvalue adjust method for \code{\link[clusterProfiler]{enrichGO}}}

\item{qvalueCutoff}{cutoff of q value}

\item{cnet.showCategory}{the number of showed cluster at cnetplot}

\item{verbose}{LuckyVerbose gseplot running message or not}

\item{save.path}{the sub path of saved files}

\item{names}{the main path and part names of saved files}
}
\description{
FastGOplot give fast way to draw GO series plot like barplot/dotplot/emapplot,Cnetplot and goplot for a enrichGO object based on clusterProfiler package.
}
\examples{
library(lucky)
need <- c("clusterProfiler","org.Hs.eg.db");Plus.library(need)
data(geneList, package = "DOSE")
data(geneList, package='DOSE')
universe <- unique(as.character(common.annot$ENTREZID))
system.time(
l <- FastGO(genes,
            geneList,
            classlevel = 2:2,
            OrgDb  = NULL,
            keyType = NULL,
            pAdjustMethod = "BH",
            pvalueCutoff = list(
               gseGO = 0.05,
               enrichGO = 0.05),
            qvalueCutoff  = 0.05,
            cnet.showCategory = 5,
            save.path = "GO",
            names = "love")
)
## enhanced plot
g <- enrichplot::cnetplot(goList$GOEnrichment$BP,
                          showCategory=10,
                          colorEdge=T,
                          node_label=T,
                          circular = F)
g1 <- g + theme(legend.text = element_text(face = "bold",size = 12))

}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
