% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_FastGSEA.R
\name{FastGSEA}
\alias{FastGSEA}
\title{Fast way to do GSEA for a geneList object}
\usage{
FastGSEA(geneList, OrgDb = NULL, keyType = NULL,
  pvalueCutoff = list(gseGO = 0.01, gseKEGG = 0.01, gseMKEGG = 0.01),
  verbose = T, save.path = "GSEA", names = "love")
}
\arguments{
\item{geneList}{a numeric with id names.For example,logFoldChang with ENSEMBL id names.Note that the names of \code{geneList} must be the same type of \code{genes}}

\item{OrgDb}{OrgDb dataset.If NULL,use "org.Hs.eg.db"}

\item{keyType}{the type of gene.Support automatically test}

\item{pvalueCutoff}{a list of pvalue adjust method for \code{\link[clusterProfiler]{gseGO}},\code{\link[clusterProfiler]{gseKEGG}} and \code{\link[clusterProfiler]{gseMKEGG}}.Note that the default value is 0.01,you can set 0.05 if you want a robust exploration.}

\item{verbose}{LuckyVerbose gseplot running message or not}

\item{save.path}{the sub path of saved files}

\item{names}{the main path and part names of saved files}
}
\description{
Fast way to do GSEA for a geneList object
}
\examples{
#'import org.Hs.eg.db
data(geneList, package = "DOSE")
gsea <- FastGSEA(geneList,
                 OrgDb = NULL,
                 keyType = NULL,
                 pvalueCutoff = list(gseGO=0.05,
                                     gseKEGG = 0.05,
                                     gseMKEGG = 0.05),
                 verbose = T,
                 save.path = "GSEA",
                 names = "love")
}
\seealso{
\code{\link[clusterProfiler]{gseGO}}; \code{\link[clusterProfiler]{gseKEGG}}; \code{\link[clusterProfiler]{gseMKEGG}}
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
