% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_FastKEGG.R
\name{FastKEGG}
\alias{FastKEGG}
\title{KEGG analysis via clusterProfiler package}
\usage{
FastKEGG(genes, geneList, default.universe = F, organism = "hsa",
  pvalueCutoff = list(enrichKEGG = 0.05, enrichMKEGG = 0.05),
  qvalueCutoff = 0.1, cnet.showCategory = 10, verbose = T,
  save.path = "KEGG", names = "love")
}
\arguments{
\item{genes}{a character of gene id.}

\item{geneList}{a numeric with id names.For example,logFoldChang with ENSEMBL id names.Note that the names of \code{geneList} must be the same type of \code{genes}}

\item{default.universe}{whether to use default universe.If \code{default.universe = F},the background genes would be provide by \code{geneList}}

\item{organism}{Species.Default is "hsa"(homo species).}

\item{pvalueCutoff}{a list of pvalue adjust method for \code{\link[clusterProfiler]{enrichKEGG}} and \code{\link[clusterProfiler]{enrichMKEGG}}}

\item{qvalueCutoff}{cutoff of q value}

\item{cnet.showCategory}{the number of showed cluster at cnetplot}

\item{verbose}{LuckyVerbose gseplot running message or not}

\item{save.path}{the sub path of saved files}

\item{names}{the main path and part names of saved files}
}
\description{
FastKEGG is a fast way to do KEGG analysis via series funtion of clusterProfiler package and automatically save related files or plots
}
\examples{
Plus.library("clusterProfiler")
data(geneList, package='DOSE')
genes <- names(geneList);
genes <- genes[1:2000]
kegg <- FastKEGG(genes = genes,
                organism = 'hsa',
                pvalueCutoff = 0.05,
                qvalueCutoff = 0.2,
                save.path = "KEGG",
                names = "love")
View(kegg)
}
