% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_FastMS.R
\name{FastMS}
\alias{FastMS}
\title{Fast Molecular Signatures Analysis via clusterProfilter}
\usage{
FastMS(genes, geneList, default.universe = T, pAdjustMethod = "BH",
  pvalueCutoff = list(enrichMS = 0.05, gseMS = 0.05),
  qvalueCutoff = 0.2, cnet.showCategory = 5, verbose = TRUE,
  save.path = "Molecular Signatures", names = "love")
}
\arguments{
\item{genes}{a character of genes.The ENSEMBL id is the best.Note that no replicate should exist.}

\item{geneList}{a named numeric of genes.The ENSEMBL id is the best.Note that no replicate should exist.}

\item{default.universe}{whether to use default universe.If \code{default.universe = F},the background genes would be provide by \code{geneList}}

\item{pAdjustMethod}{the mathod of pvalue adjustment}

\item{pvalueCutoff}{a list of pvalue adjust method for \code{\link[clusterProfiler]{enricher}} and \code{\link[clusterProfiler]{GSEA}}}

\item{qvalueCutoff}{cutoff of q value}

\item{cnet.showCategory}{the number of showed cluster at cnetplot}

\item{verbose}{LuckyVerbose gseplot running message or not}

\item{save.path}{the sub path of saved files}

\item{names}{the main path and part names of saved files}
}
\description{
Fast Molecular Signatures Analysis via clusterProfilter
}
\examples{
data(geneList, package = "DOSE")
genes = names(geneList)[1:100]
geneList = geneList
ms <- FastMS(genes,
             geneList,
             pAdjustMethod = "BH",
             pvalueCutoff = list(enrichMS=0.00001,
                                 gseMS = 0.00001),
             qvalueCutoff  = 0.2,
             cnet.showCategory = 5,
             verbose = TRUE,
             save.path = "Molecular Signatures",
             names = "love")
}
\seealso{
\code{\link[clusterProfiler]{enricher}}; \code{\link[clusterProfiler]{GSEA}}
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
