% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_FastQPCR.R
\name{FastQPCR}
\alias{FastQPCR}
\title{Fast way to deal with results of qRT-PCR}
\usage{
FastQPCR(data, sample = "samples", marker = "markers",
  bioRepeat = "biorepeat", parallelRepeat = "parepeat",
  group = "groups", group.control = "control", internal = "GAPDH",
  value = "Ct", plot.type = c(1, 2)[1], palette = NULL, size = 20,
  label.position = c(4, 5), label.type = c("p.signif", "p.format")[1],
  method = "t.test", x.title = "Genes",
  y.title = "The Relative Expression of Genes",
  legend.position = "top")
}
\arguments{
\item{data}{the result of qRT-PCR.}

\item{sample}{the colnames of sample.Note that different samples should use different sample id.}

\item{marker}{the colnames of marker}

\item{bioRepeat}{the colnames of bioRepeat.The data in "bioRepeat" would be considered as available repeat in afterward statistics.}

\item{parallelRepeat}{the colnames of parallelRepeat.The data in "parallelRepeat" would be treated via mean strategy.}

\item{group}{the colnames of group.Like "treatment" and "control".}

\item{group.control}{the names of control group.Default is "control".}

\item{internal}{the name of internal reference gene.Default is "GAPDH".}

\item{value}{the colnames of Ct value}

\item{plot.type}{one of 1 and 2.1 is used in multiple markers,but 2 is more suitable for one marker.}

\item{palette}{the color of groups.The number of color must be equal to the number of groups.Default is NULL,which mean that the strategy of lucky package would be use.}

\item{size}{ggplot parameters.the size of plot}

\item{label.position}{ggplot parameters.the position of p significance}

\item{label.type}{character string specifying label type. Allowed values include "p.signif" (shows the significance levels), "p.format" (shows the formatted p value).}

\item{method}{method use in statistics."t.test" is always recommended.You can also use "wilcox.test" for a try.Alternative choices are "anova" and "kruskal.test".}

\item{x.title}{ggplot parameters.the x axis title}

\item{y.title}{ggplot parameters.the y axis title}

\item{legend.position}{ggplot parameters.legend position}
}
\value{
a Lucky Objects
}
\description{
Fast way to deal with results of qRT-PCR,including compared plot,statistics and quality control.
}
\examples{
library(lucky)
data("qpcr",package = "lucky") # see the example of qpcr result
result.pcr <- FastQPCR(qpcr)
result.pcr <- FastQPCR(data=qpcr,
                       sample = "samples",
                       marker = "markers",
                       bioRepeat = "biorepeat",
                       parallelRepeat = "parepeat",
                       group = "groups",
                       group.control = "control",
                       internal = "GAPDH",
                       value = "Ct",
                       size=20,
                       label.y = 650,
                       method = "t.test",
                       x.title = "Genes",
                       y.title = "The Relative Expression of Genes",
                       legend.position = "top")
## Use t.test.However,it's not recommand.
result.pcr <- FastQPCR(data=qpcr,
                       sample = "samples",
                       marker = "markers",
                       bioRepeat = "biorepeat",
                       parallelRepeat = "parepeat",
                       group = "groups",
                       group.control = "control",
                       internal = "GAPDH",
                       value = "Ct",
                       size=20,
                       label.y = 650,
                       method = "t.test", #t.test
                       x.title = "Genes",
                       y.title = "The Relative Expression of Genes",
                       legend.position = "top")

## View result
View(result_qpcr$Data$statistc$whole)
View(result_qpcr$Data$statistc$pair)

## test
a = result.pcr$Data$metadata
x <- a$fc[a$markers \%in\% "marker1" & a$groups \%in\% "control"]
y <- a$fc[a$markers \%in\% "marker1" & a$groups \%in\% "treat"]
t.test(x,y)
}
\seealso{
\code{\link[ggpubr]{stat_compare_means}}
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
