% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_FastImmuScore.R
\name{FastScore}
\alias{FastScore}
\title{Fast way to calculate scores for multiple markers}
\usage{
FastScore(data, marker, time, status, cutoff, score)
}
\arguments{
\item{data}{a data frame}

\item{marker}{the colnames of markers}

\item{time}{the colname of time value for survival data}

\item{status}{the colname of status value for survival data}

\item{cutoff}{a list of cutoff for classification of each marker.See example please}

\item{score}{a list of score for classification of each marker.See example please}
}
\value{
a LuckyList
}
\description{
Fast way to calculate scores for multiple markers
}
\examples{
## This is a simulative process and available only with CORRECT VARIABLES
data <- readxl::read_xlsx("E:/RCloud/Experimental meterials/others/lijin/immu_lijin.xlsx")
data <- as.data.frame(data,stringAsFactor = F)
data$OS <- as.numeric(as.character(data$OS))
data$Osoutcome <- as.numeric(as.character(data$Osoutcome))
data$DFS <- as.numeric(as.character(data$DFS))
data$DFSoutcome <- as.numeric(as.character(data$DFSoutcome))
colnames(data)[1:4] <- c("I3","M3","I8","M8")
marker = c("I3","M3","I8","M8")
time = "OS"
status = "Osoutcome"
## 3 cutoff(n)
cutoff <- list(I3 = c(25,50,75)/100,
               M3 = c(25,50,75)/100,
               I8 = c(25,50,75)/100,
               M8 = c(25,50,75)/100)
## 4 scores(n+1)
score <- list(I3 = c(1,2,3,4),
              M3 = c(1,2,3,4),
              I8 = c(1,2,3,4),
              M8 = c(1,2,3,4))
}
\seealso{
\code{\link{FastSurvplot}}.
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
