% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_FastSurvplot.R
\name{FastSurvplot}
\alias{FastSurvplot}
\title{Fast way to draw a KM plot of ggplot style}
\usage{
FastSurvplot(data, time = "OS.time", status = "OS.status",
  marker = "pT", color = NULL, size = 16, legend = "top",
  legend.title = "T status", title = NULL, pval = TRUE,
  pval.position = c(1, 1), conf.int = F, risk.table = F,
  linetype = "solid", surv.median.line = "none", ncensor.plot = F,
  saveplot = F, name = "GC")
}
\arguments{
\item{data}{a data frame}

\item{time}{the colname of time}

\item{status}{the colname of status}

\item{marker}{the marker you want to explore with KM plot}

\item{color}{If \code{color=NULL},it would use the color from \code{mycolor} in lucky package.You can also set self-defined colors}

\item{size}{the size of plot}

\item{legend}{the position of legend}

\item{legend.title}{the title of legend}

\item{title}{the title of the KM plot}

\item{pval}{whether to output P value in KM plot}

\item{pval.position}{the position of P value in KM plot}

\item{conf.int}{whether to output confidence interval}

\item{risk.table}{whether to output risk table}

\item{linetype}{line type}

\item{surv.median.line}{whether to show median line}

\item{ncensor.plot}{whether to show ncensor plot}

\item{saveplot}{whether to save the KM plot}

\item{name}{part name of saved plot}
}
\description{
Fast way to draw a KM plot of ggplot style
}
\examples{
library(survival)
data("lung");
data = lung
data$status <- data$status -1
p <- FastSurvplot(data = data,
                  time = "time",
                  status = "status",
                  marker = "sex",
                  size = 10,
                  legend.title = "Sex",
                  pval.position = c(900,1))
p
}
\seealso{
\code{\link[survminer]{ggsurvplot}}.
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
