% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_FastWGCNA.R
\name{FastWGCNA}
\alias{FastWGCNA}
\title{Standard pipeline for WGCNA}
\usage{
FastWGCNA(expr.matrix, design, log.convert = T,
  contrast.col = "N.status", contrast.control = "N0", check = F,
  mad.portion = 0.75, mad.min = 0.01, verbose = c(goodSamplesGenes =
  3, pickSoftThreshold = 5, blockwiseModules = 3), maxBlockSize = 50000,
  minModuleSize = 30, cutoff.pval = 0.05, corType = c("pearson",
  "bicor")[1], hub_cutoffSigGM = 0.2, hub_MM = 0.8,
  hub_WeightedQ = 0.01, parallel = F, report = T, two.step = F,
  save.path = "WGCNA", names = "love")
}
\arguments{
\item{expr.matrix}{expression matrix}

\item{design}{a design object}

\item{log.convert}{whether do log2 scale for expression matrix}

\item{contrast.col}{the colname of contrast in design object like "N.status"}

\item{contrast.control}{the contrast control like "N0"}

\item{check}{whether to check critical objecta in save-file space:"dataExpr.rda","Net of WGCNA.rda","soft-thresholding list.rda".Set check=T help continue job based on the previous efforts}

\item{mad.portion}{If \code{mad.portion} <= 1, it represent the portion of data you want base on median absolute deviation(mad) filter.If \code{mad.portion} >1(for example 3000),it means the first \code{mad.portion} mad genes would be selected to use.}

\item{mad.min}{the lower cut-off of mad filter}

\item{verbose}{a named vector of verboses in multiple scenes}

\item{maxBlockSize}{set as large as possible according to the internal storage of your computer}

\item{minModuleSize}{the min count of module.Default is 30}

\item{cutoff.pval}{cut-off of the p value in significant module-phenotype filter}

\item{corType}{One of "pearson" and "bicor".Default is "pearson"}

\item{hub_cutoffSigGM}{the cut-off of significant genes-Modules in hub genes exploration.Default is 0.2}

\item{hub_MM}{the cut-off of Module Memberships in hub genes exploration.Default is 0.8}

\item{hub_WeightedQ}{the cut-off of weighted q value in hub genes exploration.Default is 0.01}

\item{parallel}{whether use enableWGCNAThreads() to accelarate cor or bicor functions}

\item{report}{whether to do a plot report}

\item{two.step}{whether use two step to complete calculation.If the nGenes is large,"two.step = T" is recommaned.}

\item{save.path}{the space of the save file.Default is "WGCNA"}

\item{names}{part of saved files name}
}
\description{
FastWGCNA give a fast and standard pipeline to do Weighted Correlation Network Analysis(WGCNA) for specified expression matrix and design object
}
\note{
1.Please choose a high-performance computer for better running. /n 2.properly set "maxBlockSize" and "mad.portion" parameters to avoid the different block produced.
}
\examples{
## This is a simulative process and NOT RUN

## data preparation
load1(c("fpkm","design.model2"))
set.seed(2018);select <- sample(1:nrow(data.fpkm),2000)
expr.matrix = data.fpkm[select,]
design = design.model2

## Quick Start
wgcna <- FastWGCNA(expr.matrix,
                  design,
                  contrast.col = "N.status",
                  contrast.control = "N0",
                  check = T,
                  parallel = F,
                  save.path = "WGCNA",
                  names = "test1");;mymusic(1)
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
\keyword{FastWGCNA}
