% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_FastWGCNA_network.R
\name{FastWGCNA_network}
\alias{FastWGCNA_network}
\title{Visualize specified module Network after FastWGCNA pipeline}
\usage{
FastWGCNA_network(object, module, hubGenes = NULL,
  hubGenes.color = mycolor[21], ratio = 0.8, size = 15,
  linkDistance = 300, linkWidth = 6, linkColour.type = NULL,
  opacityNoHover = 0, save.path = "WGCNA", names = "love")
}
\arguments{
\item{object}{the result from \code{\link{FastWGCNA}}}

\item{module}{module name.It must be a color oject like "blue"/"red"}

\item{hubGenes}{Genes you want to highlight.Default is NULL,which means no gene would be highlighted.}

\item{hubGenes.color}{the color of hubGenes.Default is mycolor[21].}

\item{ratio}{The portion of data you select}

\item{size}{a numeric to adjust the size of nodes}

\item{linkDistance}{link distance}

\item{linkWidth}{the width of link.If it is not a numeric,the default \code{"JS(function(d) { return Math.sqrt(d.value); })" } would be used.}

\item{linkColour.type}{a list of link color for \code{value} mapping.}

\item{opacityNoHover}{the legend intensity.From 0-1.Default is 0}

\item{save.path}{the space of the save file.Default is "WGCNA"}

\item{names}{part of saved files name}
}
\description{
Visualize specified module Network after FastWGCNA pipeline
}
\examples{
## NOT RUN
object = wgcna;rm(wgcna);
module = "blue"
hubGenes = c("ENSG00000009790","ENSG00000185811","ENSG00000134516","ENSG00000167613","ENSG00000122122","ENSG00000010610","ENSG00000066336","ENSG00000183813","ENSG00000102245","ENSG00000135426")

CoNet_moduleBlue <- FastWGCNA_network(object,
                                      module,
                                      hubGenes,
                                      ratio = 0.8,
                                      size = 15,
                                      linkDistance = 300,
                                      linkWidth=6,
                                      linkColour.type = NULL,
                                      opacityNoHover = 0,
                                      save.path = "WGCNA",
                                      names = "love")
}
