
\name{QC.PCA}
\alias{QC.PCA}
\title{QC.PCA}
\description{
 # QC.PCA用于绘制PCA图以观察不同组别的区分情况
}
\usage{
QC.PCA(DesignEset,
       contrast,
       savefile=F,
       names="test1")
}
\arguments{

}
\details{

}
\value{

}
\references{

}
\author{
  Weibin Huang
}
\note{

}
\seealso{

}
\examples{
QC.PCA(DesignEset,
       contrast,
       savefile=T,
       names="test1")
}
\keyword{QC.PCA}

