
\name{QC.plotMA}
\alias{QC.plotMA}
\title{QC.plotMA}
\description{
 对于给定的DesignEset和contrast可以进行MA图的绘制。
}
\usage{
QC.plotMA(DesignEset,
          contrast=NULL,
          savefile=F,
          names = "test1")
}
\arguments{
\item{DesignEset}{a DesignEset}
\item{contrast}{c("Vemurafenib","Control").对比}
\item{savefile}{F.是否将图片保存为pdf}
\item{names}{"test1".部分文件名}
}
\details{

}
\value{

}
\references{

}
\author{
  Weibin Huang
}
\note{

}
\seealso{

}
\examples{
QC.plotMA(DesignEset,
          contrast=NULL,
          savefile=F,
          names = "test1")
}
\keyword{QC.plotMA}

