
\name{QCplot}
\alias{QCplot}
\title{QCplot}
\description{
 5种QC图像函数的快速通道
}
\usage{
QCplot(DesignEset,
       contrast=NULL,
       names,
       cluster_cols=T,
       log.convert = F,
       savefile=T)
}
\arguments{

}
\details{

}
\value{

}
\references{

}
\author{
  Weibin Huang
}
\note{

}
\seealso{

}
\examples{
QCplot(DesignEset,
       contrast,
       names,
       savefile=T)
}
\keyword{QCplot}

