% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anno_get.fa.annoatation.R
\name{get.fa.annoatation}
\alias{get.fa.annoatation}
\title{create a powerful annotation object for chips containing probe suqueces information}
\usage{
get.fa.annoatation(fa.path)
}
\arguments{
\item{fa.path}{the path of fa file.It should be the .txt file from .fa via Notepad++ software.}
}
\value{
a fa.annoatation object
}
\description{
get.fa.annoatation() help get suquece associated annotation information from gencode transcripts .fa files.Before entering get.fa.annoatation(),the .fa file needs to be converted to .txt via Notpad++ software.
}
\examples{
fa.path = "E:/RCloud/database/DataDownload/annotation/unzip data/gencode.v28.transcripts.txt"
fa.annotation <- get.fa.annoatation(fa.path)
save(fa.annotation,file = "fa.annotation.rda")
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
