% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anno_get.ISSN.R
\name{get.ISSN}
\alias{get.ISSN}
\title{after get.journal.ISSN}
\usage{
get.ISSN(data, journal.name.col = "journal", ISSN.col = "ISSN.code",
  search.col = "search.names")
}
\arguments{
\item{data}{a data containing journal,ISSN.code and search.names,which is always produce by luckey::get.journal.ISSN}

\item{journal.name.col}{the colname of journals}

\item{ISSN.col}{the colname of ISSN codes}

\item{search.col}{the colname of search names}
}
\description{
get.ISSN is after get.journal.ISSN and produce unique ISSN code via intelligent algorithm.
}
\examples{
## get unique ISSN
library(lucky)
df1 <- get.ISSN(ISSN.3.Inf)
df2 <- get.ISSN2(df1)

## get IF information
data("IF")
colnames(df2)[1] <- "Full Journal Title"
library(dplyr)
ISSN_3.Inf <- left_join(df2,IF,by = "Full Journal Title")
table(is.na(ISSN_3.Inf$`Journal Impact Factor`))

## get new IF plus ISSN
colnames(ISSN_3.Inf) <- c("Journals","Search names","ISSN","Total #'Cites","IF","Eigenfactor Score")
ISSN_3.Inf <- subset(ISSN_3.Inf,select = c("Journals","ISSN","IF","Search names","Total Cites","Eigenfactor Score"))
IF_3.Inf <- ISSN_3.Inf
save(IF_3.Inf,file = "IF_3.Inf.rda")
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
