% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anno_get.ISSN2.R
\name{get.ISSN2}
\alias{get.ISSN2}
\title{After get.ISSN}
\usage{
get.ISSN2(data, journal.name.col = "journal", ISSN.col = "ISSN.code",
  search.col = "search.names")
}
\arguments{
\item{data}{a data containing journal,ISSN.code and search.names,which is always produce by luckey::get.ISSN}

\item{journal.name.col}{the colname of journals}

\item{ISSN.col}{the colname of ISSN codes}

\item{search.col}{the colname of search names}
}
\value{
data frame contain ISSN information
}
\description{
get.ISSN2 is after get.ISSN and produce more completeISSN code information.
}
\details{
Because the correction of error records need get.journal.ISSN,the speed of get.ISSN2 would be slow and please give more patience.
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
