% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anno_getAnnotation.R
\name{getAnnotation}
\alias{getAnnotation}
\title{Get annotation from a gtf annotation file}
\usage{
getAnnotation(gtf.path, select.col = NULL, Newcolnames = NULL,
  control.id = NULL)
}
\arguments{
\item{gtf.path}{the path of .gtf}

\item{select.col}{the selected colnames.Default is "gene_id","gene_name" and "gene_type"}

\item{Newcolnames}{the new colnames.Default is "ENSEMBL","SYMBOL","gene_type"}

\item{control.id}{the onlyone id colname}
}
\description{
get annotation from a gtf annotation file.
}
\examples{
getAnnotation(gtf.path,
              select.col=NULL,
              Newcolnames=NULL,
              control.id=NULL)
save(gtf.annotation,file="gtf.annotation.rda")
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
