% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anno_getAnnotation2.R
\name{getAnnotation2}
\alias{getAnnotation2}
\title{Get further annotation after getAnnotation()}
\usage{
getAnnotation2(data, geo.type = F)
}
\arguments{
\item{data}{a result from getAnnotation()}

\item{geo.type}{Default=F.If T,it considers UNIGENE as SYMBOL,which is not recommanded}
}
\value{
a data frame
}
\description{
getAnotation2()在getAnotation1()的基础上，利用org.Hs.eg.db的注释，获得类似于ENTREZID/UNIGENE等的注释信息。一般在不同的数据取交集基因时，geo.type=T有助于将UNIGENE当作SYMBOL来处理。
}
\examples{
df2.1 <- getAnnotation2(gtf.annotation,geo.type = T)
df2.2 <- getAnnotation2(gtf.annotation,geo.type = F)
}
\seealso{
\code{\link{getAnnotation}}
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
