% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_GetMEs.R
\name{GetMEs}
\alias{GetMEs}
\title{Get new module eigengenes matrix from a given dataset}
\usage{
GetMEs(object, valid.matrix, log.convert = T)
}
\arguments{
\item{object}{the result of \code{\link{FastWGCNA}}}

\item{valid.matrix}{a list of expression matrix with genes rows and sample
cols.}

\item{log.convert}{whether to do log scale for expression matrix}
}
\value{
LuckyWGCNA Object
}
\description{
Get New module eigengenes matrix from a given dataset
}
\details{
Only support wgcna object with 1 Block.
}
\examples{
library(lucky)
expr.matrix = rna.fpkm.tumor
design = rna.design.tumor
log.convert =T
newge <- GetMEs(object,expr.matrix)
View(newge$Data$MEs)
}
\seealso{
\code{\link{FastWGCNA}}.
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
\keyword{GetMEs}
