% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinic_getNomogram.R
\name{getNomogram}
\alias{getNomogram}
\title{Get nomogram,standard curves and return related data and formula}
\usage{
getNomogram(data, time.col, status.col, cluster, time.raw.type = c("Day",
  "Month", "Year")[1], time.target.type = c("Day", "Month", "Year")[2],
  time.knot = c(12, 24), dig = 9, nomo.width = 15,
  nomo.height = 12, nomo.xfrac = 0.45, standard.cmethod = c("hare",
  "KM")[2], standard.method = "boot", standard.nspot = 3,
  standard.names = "OS", standard.ylim = c(0, 1),
  standard.width = 10, standard.height = 10, names = "love1")
}
\arguments{
\item{data}{a data frame}

\item{time.col}{survival data colnames_time}

\item{status.col}{survival data colnames_status}

\item{cluster}{The vector of the marker values for which we want to compute the time-dependent ROC curves. Without loss of generality, the function assumes that larger values of the marker are associated with higher risks of events. If lower values of the marker are associated with higher risks of events, then reverse the association adding a minus to the marker values.}

\item{time.raw.type}{one of "Day","Month","Year"}

\item{time.target.type}{one of "Day","Month","Year"}

\item{time.knot}{time knot to create survival probability}

\item{dig}{Digits in nomogram plot.Default is 9.}

\item{nomo.width}{width of nomogram plot}

\item{nomo.height}{height of nomogram plot}

\item{nomo.xfrac}{saved size of nomogram plot}

\item{standard.cmethod}{cmethod of \code{\link[rms]{calibrate}}.One of 'hare' and 'KM'}

\item{standard.method}{method of \code{\link[rms]{calibrate}}.Default is 'boot'.}

\item{standard.nspot}{the number of predictive spots in standard curves.}

\item{standard.names}{type of prediction.Like "OS".}

\item{standard.ylim}{Default is c(0,1)}

\item{standard.width}{parameters of output plot}

\item{standard.height}{parameters of output plot}

\item{names}{# part of file names.}
}
\description{
Get nomogram,standard curves and return related data and formula
}
\examples{
library(lucky)
library(survival)
data(lung)
data=lung
time.col = "time"
status.col = "status"
colnames(data)
cluster = c("sex","age","ph.ecog","ph.karno") #
time.raw.type=c("Day","Month","Year")[1]
time.target.type=c("Day","Month","Year")[2]
time.knot = c(12,24)
dig=9
standard.nspot = c(4,3)
width = 15
height = 12
xfrac=0.45
names = "love1"
standard.names = "OS"
standard.ylim = c(0,1)
standard.width = 10
standard.height = 10

## draw a nomogram and output related data and formula
warnings(list1 <- getNomogram(data,
                              time.col,
                              status.col,
                              cluster,
                              time.raw.type=c("Day","Month","Year")[1],
                              time.target.type=c("Day","Month","Year")[2],
                              standard.cmethod = "KM",
                              time.knot = c(12,24),
                              standard.nspot = c(4,3),
                              standard.names = "OS",
                              names = "love1"))

#more complex example
list1 <- getNomogram(data,
                     time.col,status.col,
                     cluster,
                     time.raw.type=c("Day","Month","Year")[1],
                     time.target.type=c("Day","Month","Year")[2],
                     time.knot = c(12,24),
                     dig=9,
                     nomo.width = 15,
                     nomo.height = 12,
                     nomo.xfrac=0.45,
                     standard.nspot = 3,
                     standard.names = "OS",
                     standard.ylim = c(0,1),
                     standard.width = 10,
                     standard.height = 10,
                     names = "love1")

## calculate a time-specified survival probability
S <- list1[["formula.nomo"]]
S(25) #25 months specified survival prediction.

## other information
View(list1[["data.nomo"]])
View(list1[["data.standard"]])
}
\seealso{
\code{\link[rms]{calibrate}};\code{\link[rms]{cph}};\code{\link[rms]{nomogram}}
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
