% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anno_searchIF.R
\name{getsearch}
\alias{getsearch}
\title{get search formula with specified journal names or ISSN}
\usage{
getsearch(journal.names, searchstring = "Journal", output.names)
}
\arguments{
\item{journal.names}{a vector of journal names}

\item{output.names}{part of save file name}
}
\value{
a txt file with Pubmed search fomula
}
\description{
get search formula with specified journal names or ISSN
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
