% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinic_getTimeROC.R
\name{getTimeROC}
\alias{getTimeROC}
\title{time-dependent ROC curve base on timeROC package}
\usage{
getTimeROC(data, time.col, status.col, cause, markers,
  merge.markers = NULL, weighting.univariate = c("marginal", "cox",
  "aalen")[2], weighting.multivariate = c("marginal", "cox", "aalen")[2],
  time.raw.type = c("Day", "Month", "Year")[1],
  time.target.type = c("Day", "Month", "Year")[3], time.knot = c(3, 5),
  line.color = mycolor, size = 1.2, half.border = F,
  legend.position = c(0.8, 0.2), show.auc = T, auc.digits = 2,
  reference = T, width = 10, height = 10, merge.plot = T,
  names = "love")
}
\arguments{
\item{data}{a data frame}

\item{time.col}{time colnames}

\item{status.col}{status}

\item{cause}{the event status.For example,if 1=death,0=alive,then cause should be 1.}

\item{markers}{The vector of the marker values for which we want to compute the time-dependent ROC curves. Without loss of generality, the function assumes that larger values of the marker are associated with higher risks of events. If lower values of the marker are associated with higher risks of events, then reverse the association adding a minus to the marker values.}

\item{merge.markers}{a list of markers groups.Default is NULL.}

\item{weighting.univariate}{Default is "marginal"}

\item{weighting.multivariate}{Default is "cox"}

\item{time.raw.type}{one of "Day","Month","Year"}

\item{time.target.type}{one of "Day","Month","Year"}

\item{time.knot}{target time knots.Like 3 years and 5 years survival.}

\item{line.color}{line colors in ROC plot.}

\item{size}{the size of plot}

\item{half.border}{whether show half border of the plot}

\item{legend.position}{the position of legend position}

\item{width}{the width of plot}

\item{height}{the height of plot}

\item{merge.plot}{whether use a merge style by time.knot}

\item{names}{part of output file name.}
}
\value{
a list contain model information and series ggplot object.
}
\description{
getTimeROC is a powerful function to do time-dependent ROC curve base on timeROC::timeROC.getTimeROC help quickly and conveniently set common parameter in a ROC curve of ggplot2 style.
}
\examples{
library(lucky)
library(survival)
data(pbc)
pbc<-pbc[!is.na(pbc$trt),]
pbc$status<-as.numeric(pbc$status==2)
data = pbc
time.col = "time"
status.col ="status"
cause = 1
markers = c("bili","chol")
merge.markers=list(c("bili","chol"))
time.raw.type=c("Day","Month","Year")[1]
time.target.type=c("Day","Month","Year")[3]
time.knot=c(3,5)
names = "love"
line.color = mycolor[c(4,5,3,8,23)]
legend.position = "top"
weighting.univariate = "marginal"
weighting.multivariate = "cox"
size = 1.2
half.border=T
reference = F
width = 10
height=10
merge.plot = T
names = "love"

## merge style
models.merge <- getTimeROC(data,
                           time.col,status.col,cause,
                           markers,
                           merge.markers,
                           time.raw.type=time.raw.type,
                           time.target.type=time.target.type,
                           time.knot=time.knot,
                           line.color = line.color,
                           half.border=half.border,
                           reference = reference,
                           names = "love")

## single style
models.single <- getTimeROC(data,
                            time.col,status.col,cause,
                            markers,
                            merge.markers,
                            time.raw.type=time.raw.type,
                            time.target.type=time.target.type,
                            time.knot=time.knot,
                            line.color = line.color,
                            half.border=half.border,
                            reference = reference,
                            merge.plot = F,
                            names = "love")

## Paquid
library(timeROC);library(survival)
data(Paquid) # rm(Paquid)
Paquid$DSST <- -Paquid$DSST
Paquid$status <- ifelse(Paquid$status == 1,1,0)
models.merge <- getTimeROC(data = Paquid,
                           time.col = "time",
                           status.col = "status",
                           cause = 1,
                           markers = "DSST",
                           merge.markers = list(
                             merge = c("DSST","MMSE")
                           ),
                           time.raw.type="Year",
                           time.target.type="Year",
                           time.knot= c(3,5),
                           line.color = mycolor[c(1,3,4)],
                           names = "Paquid1")
}
\seealso{
\code{\link[timeROC]{timeROC}}.
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
