% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_GetWGCNATrait.R
\name{GetWGCNATrait}
\alias{GetWGCNATrait}
\title{create trait matrix for WGCNA}
\usage{
GetWGCNATrait(design, convert.list)
}
\arguments{
\item{design}{design object}

\item{convert.list}{a list of convert information.See examples.}
}
\value{
a new design object
}
\description{
\code{GetWGCNATrait} creates trait matrix for WGCNA.If there are some factor variables in your design object,they must be converted into numeric vector before go into \code{\link{ModuleTrait}} function,when \code{GetWGCNATrait} would complete this job.
}
\examples{
library(lucky)
data(rna.design.tumor)
stad.convert.list <- {list(
condition = list("tumor" = 1,"normal" = 0),
Mol.subtype = list("EMT" = 3,"MSI" = 2,"MSS/TP53-" = 0,"MSS/TP53+" = 1),
Mol.subtype2 = list("MSI-H" = 2,"MSI-L"=1,"MSS" = 0),
TCGA.subtype = list("CIN"=1,"EBV"=2,"GS"=3,"MSI"=4),
gender = list("FEMALE"=0,"MALE"=1),
his1 = list("Intestinal Adenocarcinoma"=0,"Adenocarcinoma" =1,"Signet Ring Type"=2),
his2 = list("Not Otherwise Specified (NOS)"=0,"Papillary Type"=1,"Tubular Type"=2,"Mucinous Type"=3,"Diffuse Type"=4),
his.grade = list("GX"=0,"G1"=1, "G2"=2, "G3"=3),
pStage = list("Stage I"=1,"Stage IA"=1,"Stage IB"=1,
             "Stage II"=2,  "Stage IIA"=2,  "Stage IIB" =2,
             "Stage III" =3,"Stage IIIA"=3, "Stage IIIB"=3,
             "Stage IIIC"=3, "Stage IV"=4),
pT = list("T1"=1, "T2"=2, "T3"=3, "T4"=4),
T.status = list("T_early"=0,"T_later"=1),
N.status = list("N0"=0, "Np"=1),
M.status  = list("M0"=0, "M1" =1,"MX"=NA)
)}
stad.Trait <- GetWGCNATrait(design,stad.convert.list)
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
\keyword{GetWGCNATrait}
