% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anno_AnnotateJournal.R
\name{MakeENlist}
\alias{MakeENlist}
\title{create ENote journal list .txt to help annotation in ENote software}
\usage{
MakeENlist(data = journals, journal.name.col = "journals",
  abbr.col = "abbr", ISSN.col = "issn", IF.col = "IF",
  IF.interval = "-", names = "love")
}
\arguments{
\item{data}{data frame or a matrix}

\item{journal.name.col}{the colname of journal full name}

\item{abbr.col}{the colname of journal abbreviation names}

\item{ISSN.col}{the colname of ISSN code}

\item{IF.col}{the colname of IF}

\item{names}{part of saved file name}
}
\description{
MakeENlist create ENote journal list .txt to help annotation in ENote software.
}
\examples{
## Use a journals data
MakeENlist(data = journals,
          journal.name.col="journals",
          abbr.col = "abbr",
          ISSN.col="issn",
          IF.col = "IF",
          IF.interval = "-",
          names = "love")
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
\keyword{MakeENlist}
