% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_ModulePreservation.R
\name{ModulePreservation}
\alias{ModulePreservation}
\title{a easy method for Module Preservation in FastWGCNA pipeline}
\usage{
ModulePreservation(object, valid.matrix = list(), nPermutations = 50,
  randomSeed = 1, quickCor = 0, verbose = 3, cutoff.Z = 10,
  report = T, width = 15, height = 7.5, save.path = "WGCNA",
  names = "love")
}
\arguments{
\item{object}{a result from \code{\link{FastWGCNA}}}

\item{valid.matrix}{a list of expression matrix with genes rows and sample
cols.}

\item{nPermutations}{specifies the number of permutations that will be
calculated in the permutation test.}

\item{randomSeed}{seed for the random number generator. If NULL, the seed
will not be set. If non-NULL and the random generator has been initialized
prior to the function call, the latter's state is saved and restored upon
exit}

\item{quickCor}{number between 0 and 1 specifying the handling of missing
data in calculation of correlation. Zero means exact but potentially slower
calculations; one means potentially faster calculations, but with
potentially inaccurate results if the proportion of missing data is large.
See cor for more details.}

\item{verbose}{integer level of verbosity. Zero means silent, higher values
make the output progressively more and more verbose.}

\item{cutoff.Z}{Default is 10.10 is efficient enought and should not be
changed.}

\item{report}{whether to do a plot report}

\item{save.path}{the space of the save file.Default is "WGCNA"}

\item{names}{part of saved files name}

\item{setLabels}{the names of expression matrix in \code{valid.matrix}.}
}
\value{
LuckyWGCNA object
}
\description{
If you want to test the preservation of modules created by
  FastWGCNA and you have also another datasets as validataion
  cohorts,\code{ModulePreservation} would be a nice function for this
  mission.
}
\details{
I.The expression matrix in \code{object} would be used as a training
  cohorts in module preservation evaluation.       II.\code{nPermutations = 200} is a usual method but it would be time-consuming.It seems that \code{nPermutations = 50} would be stable enough for module presevation exploration.
}
\examples{
## This is a simulative process and available only with CORRECT VARIABLES
object = wgcna;rm(wgcna);gc()
valid.matrix = list(Valid1 = rna.fpkm.tumor,
                    Valid2 = rna.fpkm.tumor)
result_MP <- ModulePreservation(object,
                                valid.matrix,
                                nPermutations = 20,
                                cutoff.Z = 10,
                                report = T,
                                width = 15,height = 7.5,
                                save.path = "WGCNA-test",
                                names = "love")
}
\seealso{
\code{\link{FastWGCNA}}.
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
\keyword{ModulePreservation}
