% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_ModuleSurv.R
\name{ModuleSurv}
\alias{ModuleSurv}
\title{Explore relationship between Modules and survival in FastWGCNA pipeline}
\usage{
ModuleSurv(object, design, time, status, digits = 3,
  save.path = "WGCNA", names = "love")
}
\arguments{
\item{object}{the result of  \code{\link{FastWGCNA}}.}

\item{design}{a trait-design object}

\item{time}{the time colname in design object}

\item{status}{the status colname in design object}

\item{digits}{integer indicating the number of decimal places
    (\code{round}) or significant digits (\code{signif}) to be used.
    Negative values are allowed (see \sQuote{Details}).}

\item{save.path}{the space of the save file.Default is "WGCNA"}

\item{names}{part of saved files name}
}
\value{
LuckyWGCNA object
}
\description{
ModuleSurv explores relationship between Modules and survival in FastWGCNA pipeline.
}
\details{
I.Multiple time and status is supported.    II.If object is a ME matrix,then you must make the row of ME matrix as the same as design.Pay more attention to \code{\link{ModuleSurvValid}}
}
\examples{
## This is a simulative process and available only with CORRECT VARIABLES
library(lucky)
object = wgcna;rm(wgcna);gc()
design = rna.design.tumor
time = c("OS.time","DFS.time")
status = c("OS.status","DFS.status")
result_MS <- ModuleSurv(object,
                        design,
                        time,status,
                        digits = 3,
                        height = 10,width = 8,
                        save.path = "WGCNA-test",
                        names = "love")
}
\seealso{
\code{\link{FastWGCNA}}.
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
\keyword{ModuleSurv}
