% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_ModuleSurvValid.R
\name{ModuleSurvValid}
\alias{ModuleSurvValid}
\title{Get module-Survival relationship given new validation cohorts}
\usage{
ModuleSurvValid(object, valid.matrix, log.convert = T, valid.design,
  valid.time, valid.status, digits = 3, save.path = "WGCNA",
  names = "love")
}
\arguments{
\item{object}{the result of \code{\link{FastWGCNA}}}

\item{valid.matrix}{a list of expression matrix with genes rows and sample
cols.}

\item{log.convert}{whether to do log scale for expression matrix}

\item{valid.design}{a named list of design object}

\item{valid.time}{a named list of time colnames}

\item{valid.status}{a named list of status colnames}

\item{digits}{integer indicating the number of decimal places
    (\code{round}) or significant digits (\code{signif}) to be used.
    Negative values are allowed (see \sQuote{Details}).}

\item{save.path}{the space of the save file.Default is "WGCNA"}

\item{names}{part of saved files name}
}
\value{
LuckyWGCNA object
}
\description{
Get module-Survival relationship given new validation cohorts
}
\examples{
## This is a simulative process and available only with CORRECT VARIABLES
load("E:/RCloud/RFactory/lucky/love/WGCNA-test/love_wgcna.rda");object = wgcna;rm(wgcna);gc()
 log.convert=T
valid.matrix = list(Valid1 = rna.fpkm.tumor,
                    Valid2 = rna.fpkm.tumor)
valid.design = list(Valid1 = rna.design.tumor,
                    Valid2 = rna.design.tumor)
time = c("OS.time","DFS.time")
status = c("OS.status","DFS.status")
valid.time = list(Valid1 = time,
                  Valid2 = time)
valid.status = list(Valid1 = status,
                    Valid2 = status)
save.path = "WGCNA"
names = "love"
digits = 3
result_MSV <- ModuleSurvValid(object,
                              valid.matrix,
                              log.convert=T,
                              valid.design,
                              valid.time,
                              valid.status,
                              digits = 3,
                              save.path = "WGCNA",
                              names = "love")
}
\seealso{
\code{\link{ModuleSurv}}.
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
\keyword{ModuleSurvValid}
