% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_ModuleTrait.R
\name{ModuleTrait}
\alias{ModuleTrait}
\title{Enhanced Module-Trait Relationships Plot after FastWGCNA Process}
\usage{
ModuleTrait(object, design, variable, corType = c("pearson", "bicor")[1],
  report = F, height = 10, width = 8, save.path = "WGCNA",
  names = "love")
}
\arguments{
\item{object}{the result of  \code{\link{FastWGCNA}}.}

\item{design}{a trait-design object}

\item{variable}{the variables you want to show in Module-Trait relationships plot.}

\item{corType}{one of "pearson" and "bicor".Default is pearson}

\item{report}{whether do a plot report}

\item{height, width}{the size of saved plot}

\item{save.path}{the space of the save file.Default is "WGCNA"}

\item{names}{part of saved files name}
}
\value{
a Module-Trait relationships plot
}
\description{
After \code{\link{FastWGCNA}}, you can input the result into \code{ModuleTrait} function,and get a more powerful Module-Trait relationships visualization.
}
\details{
trait-design object means the design should convert to a numeric one.A good method is to use \code{\link{GetWGCNATrait}}.
}
\examples{
## This is a simulative process and NOT RUN
library(lucky)
object = wgcna;rm(wgcna);gc()
design = rna.design.tumor
variable = c("age","his1","gender","N.status","T.status")
save.path = "WGCNA"
names = "love"
corType = "pearson"
height = 10;width = 8
result_MT <- ModuleTrait(object,
                         design,
                         variable,
                         corType = "pearson",
                         report=T,
                         height = 10,width = 8,
                         save.path = "WGCNA",
                         names = "love")
}
\seealso{
\code{\link{GetWGCNATrait}}
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
\keyword{ModuleTrait}
