% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_ModuleTraitValid.R
\name{ModuleTraitValid}
\alias{ModuleTraitValid}
\title{Get module-trait relationship given new validation cohorts}
\usage{
ModuleTraitValid(object, valid.matrix, log.convert = T, valid.design,
  valid.variable, corType = c("pearson", "bicor")[1], report = T,
  height = 10, width = 8, save.path = "WGCNA", names = "love")
}
\arguments{
\item{object}{the result of \code{\link{FastWGCNA}}}

\item{valid.matrix}{a list of expression matrix with genes rows and sample
cols.}

\item{log.convert}{whether to do log scale for expression matrix}

\item{valid.design}{a named list of design object}

\item{valid.variable}{a named list of trait vector}

\item{corType}{one of "pearson" and "bicor".Default is pearson}

\item{report}{whether do a plot report}

\item{height}{the size of saved plot}

\item{width}{the size of saved plot}

\item{save.path}{the space of the save file.Default is "WGCNA"}

\item{names}{part of saved files name}
}
\value{
LuckyWGCNA object
}
\description{
Get module-trait relationship given new validation cohorts. \code{ModuleTraitValid} supports multiple expression matrix.
}
\examples{
## This is a simulative process and available only with CORRECT VARIABLES
load("E:/RCloud/RFactory/lucky/love/WGCNA-test/love_wgcna.rda")
object = wgcna;rm(wgcna);gc()
log.convert=T
valid.matrix = list(Valid1 = rna.fpkm.tumor,
                    Valid2 = rna.fpkm.tumor)
valid.design = list(Valid1 = rna.design.tumor,
                    Valid2 = rna.design.tumor)
variable = c("age","his1","gender","N.status","T.status")
valid.variable = list(Valid1 = variable,
                      Valid2 = variable)
corType = c("pearson","bicor")[1]
report=F
height = 10;width = 8
save.path = "WGCNA"
names = "love"
result_MTV <- ModuleTraitValid(object,
                               valid.matrix,log.convert=T,
                               valid.design,
                               valid.variable,
                               corType = c("pearson","bicor")[1],
                               height = 10,width = 8,
                               save.path = "WGCNA",
                               names = "love")
}
\seealso{
\code{\link{ModuleTrait}}.
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
\keyword{ModuleTraitValid}
