% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_environment.R
\name{mymusic}
\alias{mymusic}
\title{play music via tuneR}
\usage{
mymusic(n = 1, path = NULL)
}
\arguments{
\item{n}{the order of available music paths}

\item{path}{the path of music files ducument}
}
\value{
a string of music file path
}
\description{
play music via tuneR
}
\examples{
mymusic(1)
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
\keyword{mymusic}
