% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_Plus.library.R
\name{Plus.library}
\alias{Plus.library}
\title{Automatically install and library package}
\usage{
Plus.library(packages)
}
\arguments{
\item{packages}{a string vector of package names}
}
\description{
A fast way for Chinese people to automatically install and library package.
}
\examples{
Plus.library(c("DT","RColorBrewer","stringr"))

}
\author{
Weibin Huang\email{654751191@qq.com}
}
