% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anno_ProteinAnnotation.R
\name{ProteinAnnotation}
\alias{ProteinAnnotation}
\title{Get protein annotation from gpl file}
\usage{
ProteinAnnotation(gtf.path, gene.col = "gene_id",
  protein.col = "protein_id")
}
\arguments{
\item{gtf.path}{the path of .gtf}

\item{gene.col}{the colname of gene_id(ENSG)}

\item{protein.col}{the colname of protein_id(ENSP)}
}
\value{
a data frame of annotation for protein(ENSP id)
}
\description{
Get protein annotation from gpl file
}
\examples{
gtf.path='E:/RCloud/database/DataDownload/annotation/unzip data/gencode.v28.chr_patch_hapl_scaff.annotation.gtf'
ProteinAnnotation(gtf.path)
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
\keyword{ProteinAnnotation}
