% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_rond.plus.R
\name{round2}
\alias{round2}
\title{publish-level round}
\usage{
round2(x, digits = 3)
}
\arguments{
\item{x}{a numeric vector}

\item{digits}{decimal places}
}
\description{
round.plus is a specific round method for print decimals in publication.round2 is based on lucky::round.plus.round2 is recommanded to do an optimized show.
}
\details{
round2 would not deal with value like NA and Inf.round2 supports multiple vector.
}
\examples{
round2(200.4526)
round2(1)
round2(0.9)
round2(0.00258)
round2(0.002)
round2(0.001)
round2(0.0004)
a <- c(200.4526,1,0.9,0.00258,0.002,0.001,0.0004,NA,Inf,-Inf)
round2(a)
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
\keyword{round2}
