% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_ScreenGenes.R
\name{ScreenGenes}
\alias{ScreenGenes}
\title{DEGs screening via common t-test}
\usage{
ScreenGenes(object, design, contrast.col, contrast.level = c("tumor",
  "normal"), contrast.control = c("normal"), cutoff.p = 0.05,
  verbose = F)
}
\arguments{
\item{object}{DesignList or expression matrix}

\item{design}{design object.If \code{object} is DesignList,it can be NULL}

\item{contrast.col}{the colnames of contrast.Like "N.status"}

\item{verbose}{whether report the process}
}
\value{
data frame of DEGs information
}
\description{
DEGs screening via common t-test
}
\examples{
## This is a simulative process and available only with CORRECT VARIABLES
res <- ScreenGenes(DesignEset,
                   contrast.level = c("tumor","normal"),
                   contrast.control = c("normal"),
                   cutoff.p=0.05)
}
\seealso{
\code{\link{DESeq1}}.
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
