% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anno_searchIF.R
\name{SearchIF2}
\alias{SearchIF2}
\title{SearchIF2:get journal list with specified strings of journal names}
\usage{
SearchIF2(data, col.journal.names = "Full Journal Title",
  col.IF = "Journal Impact Factor", string = "Nature", lower.limit)
}
\arguments{
\item{data}{IF.data.a data frame containing IF and journal names information}

\item{col.journal.names}{the colnames of journal names}

\item{col.IF}{the colnames of IF}

\item{string}{part of journal names.Always a string like "nature" or "cell"}

\item{lower.limit}{the lower CI of IF in selected journals}
}
\value{
a txt file with Pubmed search fomula and a csv with journal names and IF.
}
\description{
get journal list with specified strings of journal names
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
