% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anno_SeqmanAfter.R
\name{SeqmanAfter}
\alias{SeqmanAfter}
\title{extra information from match_probe.txt from Seqman software}
\usage{
SeqmanAfter(path.match.probe, information.col = "trans_id",
  save.file = T, names = "test1")
}
\arguments{
\item{path.match.probe}{the match_probe file created by Seqman software from Sangerbox.}

\item{information.col}{annotation information colname}

\item{save.file}{whether save results as rda}

\item{names}{part of saved file name.}
}
\value{
a data frame containing chip annotation information.
}
\description{
In chip annotation,after Seqman process,there is a file with patterns like "match_probe",and SeqmanAfter() help to extra information from it.
}
\examples{
path.match.probe <- "E:/RCloud/database/DataDownload/GEOqueryDownload/GSE17154/GPL6699_match_probe.txt"
information.col="trans_id"
SeqmanAfter(path.match.probe,
            information.col,
            save.file=T,
            names ="GPL6699")
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
