% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anno_SeqmanBefore.R
\name{SeqmanBefore}
\alias{SeqmanBefore}
\title{prepare chip probe information(.txt) before seqman pipelin}
\usage{
SeqmanBefore(eset, probe.id.col = "ID", sequence.col = "SEQUENCE",
  save.file = T, names = "test1")
}
\arguments{
\item{eset}{an eset object containing feature data.}

\item{probe.id.col}{the colname of the probe id}

\item{sequence.col}{the colname of the sequence information}

\item{save.file}{whether save results as rda}

\item{names}{part of saved file name}
}
\value{
a .txt file containing probe ids and sequeces.
}
\description{
When we want to annotate a chip with probe sequence,Seqman software is a good choice.Before use Seqman,a probe information should be prepared,which SeqmanBefore() can help do.
}
\examples{
load("E:/RCloud/database/DataDownload/GEOqueryDownload/GSE17154/GSE17154_rawEset.rda")
probe.id.col="ID"
sequence.col = "SEQUENCE"
test1 = SeqmanBefore(eset,
                     probe.id.col,
                     sequence.col,
                     save.file=T,
                     names = "test1")
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
