% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_SequenceMotif.R
\name{SequenceMotif}
\alias{SequenceMotif}
\title{co-motif exploration for sequence}
\usage{
SequenceMotif(sequence, motif, base = c("A", "C", "U", "G"),
  verbose = T)
}
\arguments{
\item{sequence}{a vector of sequences}

\item{motif}{a charactor or numeric motif}

\item{base}{the type of base.Default is \code{c("A","C","U","G")}}

\item{verbose}{whether do reports}
}
\value{
a list contain counts and sequences information of motifs
}
\description{
co-motif exploration for sequence
}
\examples{
## Create some random sequences
sequence <-  NULL
for(i in 1:100){
  set.seed(i);
  j <- sample(18:21,1)
  s.i <- sample(c("A","C","U","G"),j,replace = T)
  s.i <- paste(s.i,collapse = "")
  sequence <- c(sequence,s.i)
}

## explore two 4-base motifs
motif <- c("CCGU","GGAU")
l <- SequenceMotif(sequence,motif)
View(l$Counts)
View(l$Sequence)
l$Sequence$GGAU

## explore all 3-base or 4-base motifs
motif <- c(3,4)
l <- SequenceMotif(sequence,motif)
View(l$Counts)
View(l$Sequence)
l$Sequence$GGAU
}
\seealso{
\code{\link{combn2}};\code{\link[plyr]{alply}};
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
