% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_starBase1.R
\name{starBase1}
\alias{starBase1}
\title{Data Clean for starBase}
\usage{
starBase1(path, mode = "RNA-RNA")
}
\arguments{
\item{path}{the absolute path of seriese file(.xls) download from starbase}

\item{mode}{Only available for "RNA-RNA".Other is on schedule}
}
\value{
a data frame
}
\description{
Data Clean for starBase
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
\keyword{straBase1}
