% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_Survplot.genedata.R
\name{Survplot_genedata}
\alias{Survplot_genedata}
\title{Survival K-M plot base on expression matrix and design object}
\usage{
Survplot_genedata(expr.matrix, design, select, event.status, event.time,
  event.lower = 0, mode = c("median", "maxstat")[2],
  pval.position = NULL, size = 12, save.file = F, show.music = F,
  width = 12, height = 12, names = "love")
}
\arguments{
\item{expr.matrix}{expression matrix}

\item{design}{design object}

\item{select}{the genes or markers you want to select}

\item{event.status}{the colname of the status of event.}

\item{event.lower}{the lower cutoff of event time.Default is 0.}

\item{mode}{the mode of cut-off strategy for continuous value like gene expression data.Default is \code{mode = "maxstat"} provided by \code{\link[maxstat]{maxstat.test}}.If you want median value as cut-off,you can just set \code{mode = "median"}}

\item{pval.position}{the postion of p value in the plot}

\item{size}{the size of saved plot}

\item{save.file}{whether to save pdf for plot.If you select more than 1 genes,it would set \code{save.file = T} automatically.}

\item{show.music}{whether to show music when job was completed.Only available when you select more than 1 genes.}

\item{width}{the width of saved plot}

\item{height}{the height of saved plot}

\item{event.timethe}{colname of the time of event.}

\item{part}{name of saved pdf file}
}
\description{
Survplot.genedata help draw  KM plot base on expression matrix and design object
}
\details{
There are some different between the output of multi- and uni- variate plot.
}
\examples{
library(lucky);
data("rna.tpm",package = "lucky")
data("rna.design",package = "lucky")

## Muti-variable Plot.here I focus on patients with more follow-up time than 89 days with event.lower = 89
p <- Survplot_genedata(expr.matrix = log2(rna.tpm + 1),
                       design = rna.design,
                       select = c("ENSG00000004478",
                                  "ENSG00000000457"),
                       event.status = "OS.status",
                       event.time = "OS.time",
                       event.lower = 89,
                       mode = c("median","maxstat")[2],
                       pval.position = c(900,1),# x=900,y=1
                       size = 12,
                       save.file = F,
                       show.music = F,
                       width = 12,height = 12,
                       names = "love")

## univariable Plot
p <- Survplot_genedata(expr.matrix = log2(rna.tpm + 1),
                       design = rna.design,
                       select = "ENSG00000004478",
                       event.status = "OS.status",
                       event.time = "OS.time",
                       event.lower = 89,
                       mode = c("median","maxstat")[2],
                       pval.position = c(900,1))
}
\seealso{
\code{\link[maxstat]{maxstat.test}}
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
