% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BI_testBatchEffect.R
\name{testBatchEffect}
\alias{testBatchEffect}
\title{test and remove potencial batch effects}
\usage{
testBatchEffect(SigObject, design = NULL, candidate = c("gender",
  "diognosis.year"), n.sv = NULL, names = "love", report = T)
}
\arguments{
\item{SigObject}{the result of \code{\link{DESeq1}}, \code{\link{edgeR1}} or a expression matrix}

\item{design}{design object.If \code{SigObject} is a matrix,it is needed.}

\item{candidate}{the candidate variables you selectd like years or days}

\item{names}{the name of saved file}

\item{report}{whether to show plot reports}
}
\description{
testBatchEffect help test potencial and specified fators that may bring batch effect base on DESeq2 pipeline
}
\examples{
## This is a simulative process and available only with CORRECT VARIABLES
coldata <- as.data.frame(colData(SigObject$dds))
colnames(coldata)
candidate=c("gender","diognosis.year","pT")
t.be <- testBatchEffect(SigObject,
                        candidate,
                        names = "test")
View(t.be)
View(t.be$Unknown$SigTest)
}
\seealso{
\code{\link{DESeq1}}.
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
\keyword{testBatchEffect}
