% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Supplement_testFastBar.R
\name{testFastBar}
\alias{testFastBar}
\title{Fast Way to Get a Bar plot based on ggpubr}
\usage{
testFastBar(data, x, y, order = NULL, fill = "white", color = NULL,
  palette = NULL, label = TRUE, size = 15,
  position = position_stack(), title = "Frequecy",
  legend.position = "right")
}
\arguments{
\item{data}{a data frame}

\item{x}{x is usually the colnames of symbol/names}

\item{y}{y is usually the colnames of frequency}

\item{order}{whether order from upper to lower}

\item{fill}{Default is white.Set fill=x is a useful way to enhance visualize barplot and recommanded}

\item{color}{Default is NULL,and use x as color scale}

\item{palette}{Default is NULL,and it use a lucky::mycolor color strategy.Or you can set self-defined colors.Note that the number of color you want to set should be the same as the unique number of x}

\item{label}{Default is TRUE.If variables are too much,you can set FALSE to enhance visualization}

\item{size}{a size control for plot title,axis titles and labels}

\item{position}{see ggpubr::ggbarplot}

\item{title}{the title of the plot}

\item{legend.position}{the position of legend.you can select "top","left","right" and "none"}
}
\value{
a ggplot list
}
\description{
Fastbar help get a barplot of ggplot2 style.
}
\examples{
## This is a simulative process and NOT RUN
p  <- FastBar(data = df,
              x = "SYMBOL",y = "Freq",
              fill = "SYMBOL",
              title = "Frequecy",
              legend.position = "none",
              size = 20)
}
\seealso{
ggpubr::ggbarplot
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
\keyword{FastBar}
