% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_testPrimer.R
\name{testPrimer}
\alias{testPrimer}
\title{Give Mutiple primers,test the amplication sequence and whether cross different exons}
\usage{
testPrimer(forward, reverse, path.transcript.fa = NULL,
  path.exon.fa = NULL, verbose = T)
}
\arguments{
\item{forward}{a name vector of forward primers}

\item{reverse}{a name vector of reverse primers}

\item{path.transcript.fa}{the path of transcript(cDNA) fa file.if primers are multiple,it is recommanded to set \code{path.transcript.fa=NULL}}

\item{path.exon.fa}{the path of exons fa file.if primers are multiple,it is recommanded to set \code{path.exon.fa=NULL}}
}
\value{
A luckyList
}
\description{
Give Mutiple primers,test the amplication sequence and whether cross different exons
}
\examples{
library(lucky)
forward = c(DSCR8="GGGTGGCAAAAAGAGAAGATGG",
            DSCR8 = "CGTCTTGTCCACTCGTTTCTG",
            DSCR8 = "AAATAGGCTTCCCACTTGGC",
            DSCR8 = "GGGTGGCAAAAAGAGAAGATGG")
reverse = c(DSCR8="GGTCCAGGCTCCTTCATTTTTG",
            DSCR8 = "GGCATGTTCTAAATGGCAGCTC",
            DSCR8 = "CCAGGCTCCTTCATTTTTGCTC",
            DSCR8 = "GGTCCAGGCTCCTTCATTTTTG")
res <- testPrimer(forward,reverse)

}
\seealso{
\code{\link[Biostrings]{XStringSet-io}};\code{\link[XVector]{XVector-class}};\code{\link[Biostrings]{matchPattern}};\code{\link{testPrimer1}}
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
