% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_testPrimer.R
\name{testPrimer1}
\alias{testPrimer1}
\title{Give a primer,test the amplication sequence and whether cross different exons}
\usage{
testPrimer1(forward, reverse, path.transcript.fa, path.exon.fa,
  verbose = T)
}
\arguments{
\item{forward}{forward primer}

\item{reverse}{reverse primer}

\item{path.transcript.fa}{the path of transcript(cDNA) fa file}

\item{path.exon.fa}{the path of exons fa file}
}
\value{
A luckyList
}
\description{
Give a primer,test the amplication sequence and whether cross different exons
}
\examples{
## This is a simulative process and available only with CORRECT VARIABLES
forward = "AAGGAGCCTGGACCCAACTT"
reverse = "TGAGCCAGCCTTCATTTTCCA"
path.transcipt.fa = "E:/iProjects/XYJ.lnc/constant/Homo_sapiens_DSCR8_cDNA.fa"
path.exon.fa = "E:/iProjects/XYJ.lnc/constant/Homo_sapiens_DSCR8_Exons.fa"

res <- testPrimer1(forward,
                   reverse,
                   path.transcript.fa,
                   path.exon.fa)
## see result
names(res$Data$Amplication)
res$Data$CrossExons

}
\seealso{
\code{\link[Biostrings]{XStringSet-io}};\code{\link[XVector]{XVector-class}};\code{\link[Biostrings]{matchPattern}};
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
