% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_WesternBlotquantity.R
\name{WesternBlotQuantity}
\alias{WesternBlotQuantity}
\title{Tool for protein quantity in Western Blot}
\usage{
WesternBlotQuantity(name = NULL, path = NULL)
}
\arguments{
\item{name}{the name of saved .rda}

\item{path}{Default is NULL.You have to set up an available.Please see the example file in lucky package}
}
\description{
Tool for protein quantity in Western Blot
}
\examples{
## See the example data
example <- system.file("extdata", "wb quantity.xlsx", package = "lucky")
View(readxl::read_xlsx(example))

## A common and useful running
res <- WesternBlotQuantity()
View(res$Data$metadata)
View(res$Data$result)

## If you want a backup,just set the 'name' parameter.The .rda file would be saved in the specified path.
res <- WesternBlotQuantity("I love you")
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
