% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anno_AnnotateJournal.R
\name{AnnotateJournal}
\alias{AnnotateJournal}
\title{annotate journal via letpub}
\usage{
AnnotateJournal(journal.id = 1:13000, source = "letpub", report = T)
}
\arguments{
\item{journal.id}{a vector of journal id}

\item{source}{the data source.Default is letpub.}

\item{report}{whether report working information}
}
\description{
AnnotateJournal help annotate journal id via letpub website.AnnotateJournal is more powerful than get.ISSN series and is recommanded to get information of journals like ISSN or IF.
}
\examples{
AnnotateJournal(journal.id=1:13000,
                source = "letpub")
}
\author{
Weibin Huang<\email{654751191@qq.com}>
}
\keyword{AnnotateJournal}
